==> Step4_Calibration.Rmd <==
---
title: "RHESSysIOinR_Coweeta_WS32_74 Multi Site Validation Field Data_Step4"
author: "Carlos Quintero"
date: "6/17/2024"
output:
  html_document: default
  pdf_document: default
---
```{r include=FALSE}
knitr::opts_knit$set(root.dir = system.file("extdata/", package = "RHESSysIOinR"))
 setwd(system.file("extdata/", package = "RHESSysIOinR"))
```

Coweeta Hydrologic Laboratory Watershed 32 using Static, SSURGO, and RSS soil inputs. 

The following R Markdown script prepares and runs the Regional Hydro-Ecologic Simulation System (RHESSYS) within the R environment. 
Modified from RHESSysIOinR example scripts created by Will Burke and Ryan Bart. Includes CLHS code snippets and SDA code from Dylan Beaudette.

Windows Subsystem via Linux must be installed before RHESSys will run in a Windows environment.
GRASS8 must be installed to run the spatial preprocessing portion of the script.

Code has been modified to run RHESSys 7.4 and 5.18. 
7.4 vegetation processing differs slightly from 5.18.
Various template differences exist between the two versions of RHESSys and template files must be carefully reviewed if moving between versions.
This script is currently set up to run RHESSys 7.4

```{r setup, include=TRUE, verbose = FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
#install.packages("devtools")
#devtools::install_github("ropensci/FedData")

library(aqp)
library(daymetr)
library(clhs)
library(EcoHydRology)
library(elevatr)
library(ezknitr)
#library(FedData)
library(foreach)
library(foreign)
library(ggplot2)
library(ggpubr)
library(hydroGOF)
library(imputeTS)
library(knitr)
library(latticeExtra)
library(lubridate)
library(Metrics)
library(raster)
library(rasterVis)
library(readxl)
library(reshape2)
library(rgdal)
library(rgrass)
library(rmarkdown)
library(Rmisc)
library(RHESSysPreprocessing)
library(RHESSysIOinR)
library(sf)
library(soilDB)
library(sp)
library(stats)
library(stringi)
library(tactile)
library(terra)
library(testthat)
library(tidyverse)
library(viridisLite)
library(dplyr)

setwd(system.file("extdata/", package = "RHESSysIOinR"))
getwd()
rh_ver = dir(path = "modelfiles/rhessys/", pattern = "^rhessys\\d+",recursive = F)
rh_path = file.path("modelfiles/rhessys", rh_ver)

## Load in necessary files
filedir <- ("C:/Users/Carlos/Documents/GitHub/RHESSys74-R-Markdown/")

watershedshapefile <- paste0(filedir, "input_files/shapefiles/watersheds/Coweeta_Hydrologic_Laboratory.shp")
streamshapefile<- paste0(filedir,"input_files/shapefiles/streams/StreamsCaldwellClipped.shp")
weirshapefile<- paste0(filedir,"input_files/shapefiles/weirs/Subbasinoutlets.shp")
roadshapefile <- paste0(filedir, "input_files/shapefiles/roads/UpdatedRoads4Reprojected32617.shp")

CWRG12 <- paste0(filedir,"input_files/precipitation/RDS-2017-0031/Data/RG12_daily_1942_2021.csv")
CWRG20 <-paste0(filedir,"input_files/precipitation/RDS-2017-0031/Data/RG20_daily_1962_2021.csv")

CS21<- paste0(filedir,"input_files/temperature/RDS-2015-0042/Data/cs21_daily.csv")
CS28<- paste0(filedir,"input_files/temperature/RDS-2015-0042/Data/cs28_daily.csv")

newsmlocations<-  paste0(filedir,"input_files/shapefiles/soil_moisture/CoweetaNRCSsmKML.kml")

```
New Patch Filter

Basin - Basin 
Hillslope - Subbasin
Zone - Patch
Patch - Patch 

1 112 136557
1 110 141942
1 110 149478


## note that streamflow output from patch filter is in mm
```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

patchfilterpatch1 = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "ws32patch1",
                                   spatial_level = "patch",
                                   spatial_ID = "1:69:144625:144625",
                                   variables = c("rootzone.S","rootzone.potential_sat","rootzone.field_capacity", "rz_storage"))

patchfilterpatch2 = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "ws32patch2",
                                   spatial_level = "patch",
                                   spatial_ID = "1:69:149472:149472",
                                   variables = c("rootzone.S","rootzone.potential_sat","rootzone.field_capacity", "rz_storage"))

patchfilterpatch3 = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "ws32patch3",
                                   spatial_level = "patch",
                                   spatial_ID = "1:69:157007:157007",
                                   variables = c("rootzone.S","rootzone.potential_sat","rootzone.field_capacity", "rz_storage"))


basinfilter_static = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "static_basin_daily",
                                   spatial_level = "basin",
                                   spatial_ID = as.integer("1"),
                                   variables = c("hill.gw.Qout","patch.rootzone.S","patch.rootzone.potential_sat","patch.rootzone.field_capacity", "patch.rz_storage","patch.total_water_in", "patch.streamflow", "patch.evaporation" ))


basinfilter_ssurgo = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "ssurgo_basin_daily",
                                   spatial_level = "basin",
                                   spatial_ID = as.integer("1"),
                                   variables = c("hill.gw.Qout","patch.rootzone.S","patch.rootzone.potential_sat","patch.rootzone.field_capacity", "patch.rz_storage","patch.total_water_in", "patch.streamflow", "patch.evaporation" ))


basinfilter_dsm = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "dsm_basin_daily",
                                   spatial_level = "basin",
                                   spatial_ID = as.integer("1"),
                                   variables = c("hill.gw.Qout","patch.rootzone.S","patch.rootzone.potential_sat","patch.rootzone.field_capacity", "patch.rz_storage","patch.total_water_in", "patch.streamflow", "patch.evaporation" ))

# start_time = Sys.time()
# start_time
# run_rhessys_single(input_rhessys = input_rhessys,
#                    hdr_files = input_hdr,
#                    tec_data = input_tec_data,
#                    cmd_pars = stdpars,
#                    output_filter = c(patchfilterpatch1,patchfilterpatch2,patchfilterpatch3),
#                    runID = 'single')
# end_time = Sys.time()
# end_time - start_time
# end_time

```

Setup RHESSys Calibration, CLHS used to reduce amount of samples processed

The model calibrates over the following parameters.

M - Decay of hydraulic conductivity with depth
K - Hydraulic conductivity at the surface
Sd - soil depth
Mv - Vertical decay of hydraulic conductivity with depth
Kv - Vertical hydraulic conductivity at the surface
GW1 - Saturated to Groundwater Coefficient
GW2 - Groundwater Loss Coefficient

Parameter & Typical Range from RHESSys Wiki

M - .01 - 20
K - 1 - 600
GW1 -  .001 - .3
GW2 -  .01 - .9

" In order to generate an adequate sample across the full range for each parameter, values from .001 - .01, values from .01 - 1, and values over 1 should be generated separately. The m and K parameters are multipliers on the initial values set for m and K (values assigned to the m and K maps)." - Taken from RHESSys Wiki

Additional Calibration Parameters are available natively although they are not as well documented

pa - multiplier to scale the pore size index
po - multiplier to scale the psi air entry 

vseng1 - multiplies specific leaf areas
vseng2 - multiplies the ratio of shaded to sunlit leaf area
vseng3 - Multiplier used only with the Dickenson algorithm of carbon allocation (set with the epc.allocation_flag variable in the vegetation definition file). It changes the allocation of net photosynthate sensitivity based on the current LAI. If not using the Dickenson strategy of carbon allocation (i.e. using Waring or default Constant strategies), set third value to 1.0. (i.e. -vgsen 1.0 2.0 1.0)

```{r RHESSys Calibration Parameter Setup}
setwd(system.file("extdata/", package = "RHESSysIOinR"))
# number of samples for full data space
fs <- 10000
# number of samples to calibrate over, 100 samples take approx 40 mins in base R and 1 hour in rmarkdown
n <- 1000
# create data frame
d <- data.frame(
  m = runif(fs, min=0.1, max=20),
  k = runif(fs, min=0.5, max=200),
  soil_dep = runif(fs, min=0.1, max=5),
  m_v = runif(fs, min=0.1, max=20),
  k_v = runif(fs, min=0.1, max=20),
  gw1 = runif(fs, min=0.01, max=0.3),
  gw2 = runif(fs, min=0.01, max=0.7),
  pa = runif(fs, min=0.1, max=3.0),
  po = runif(fs, min=0.1, max=3.0),
  vgseng1 = runif(fs, min=0.1, max=3.0),
  vgseng2 = runif(fs, min=0.1, max=3.0),
  vgseng3 = runif(fs, min=0.1, max=3.0))
# marginal distributions, for analysis
m <- melt(d)
bwplot(variable ~ value, data = m, par.settings = tactile.theme(), main = "distribution of calibration parameters before clhs subsetting")
# cLHS subset
s <- clhs(d, size = n, simple = TRUE, use.cpp = TRUE)
# combine full dataset + subset
g <- make.groups(full = d, cLHS = d[s, ])
# create new list
params<- d[s, ]

params[1:5,]

## Quick and dirty way to include "uncalibrated" result to test against, groundwater model still included for now, unsure of what to do with parameters for gw
## params[1,] = c(1,1,1,1,1,0,0)
## did not produce significantly different result

#write param table to specific text file if file does not exist
## if name is not changed then old table is used
# paramtablename = "out/cwparamtable.txt" 

getwd()

##UNCOMMENT TO WRITE NEW PARAMS
#  write.table(params, file = "out/cwparamtable.txt")
###

### writes new params table with todays date to a text file
#write.table(params, file = paste("out/cwparamtable",format(Sys.Date(),"%m%d%y"),".txt",sep = ""))

## reads in specific params file, used for keeping input params constant for analysis
## do not rewrite file if it exits

#{if (file.exists("out/cwparamtable.txt"))
#   print("text")}

params <- read.table(file = "out/cwparamtable.txt")

# prepare RHESSys for run
params<- as.data.frame(params)

input_rhessys = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32dsm.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32dsm",
  flowtable = "CWWS32dsm.flow",
  start = "2013 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32",
  commandline_options = c("-b"))

#commandline_options = c("-b -p 1 128 154914 154914 ")) <- Patch Specific Output
  
input_tec_data = IOin_tec_std(start = "2015 11 1 1",
                              end = "2018 11 1 1",
                              output_state = FALSE)

soildefs <- list.files("defs/", pattern = "soil_", full.names = TRUE)

input_hdr = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = c("defs/lu_undev.def","defs/lu_agriculture.def","defs/lu_raingarden.def","defs/lu_urban.def"),
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")
 
```
 
Run RHESSys n times sequentially

# NO LONGER USED

```{r RHESSys Sequential Calibration, eval=FALSE, include=FALSE}
start_time = Sys.time()

for(i in 1:n) {
 
stdpars<- IOin_cmd_pars(
  m = params[i,]$m,
  k = params[i,]$k,
  soil_dep= params[i,]$soil_dep,
  m_v = params[i,]$m_v,
  k_v = params[i,]$k_v,
  gw1 = params[i,]$gw1,
  gw2 = params[i,]$gw2,
  pa = params[i,]$pa,
  po = params[i,]$po,
  vgseng1 = params[i,]$vgseng1,
  vgseng2 = params[i,]$vgseng2,
  vgseng3 = params[i,]$vgseng3)

run_rhessys_single(input_rhessys = input_rhessys,
                   hdr_files = input_hdr,
                   tec_data = input_tec_data,
                   cmd_pars = stdpars,
                   runID = i)} 

end_time = Sys.time()

end_time - start_time

```

Run RHESSys in Parallel using the foreach package, leaving 4 cores available for other tasks

Takes about 1 hour for 100 runs at 10x10m Daily WS32.

# NO LONGER USED
# Modify for MCMC
## must refine calibration to do multiple steps, first gw1 gw2, then other variables
## multi point calibration and validation
```{r Run RHESSYS in Parallel using foreach, eval=FALSE, include=FALSE}
n.cores <- parallel::detectCores() - 4  
  
my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
doParallel::registerDoParallel(cl = my.cluster)

start_time = Sys.time()
  
  foreach(i = 1:n, .packages = 'RHESSysIOinR') %dopar% {
    
    stdpars<- IOin_cmd_pars(m = params[i,]$m,
                            k = params[i,]$k,
                            soil_dep= params[i,]$soil_dep,
                            m_v = params[i,]$m_v,
                            k_v = params[i,]$k_v,
                            gw1 = params[i,]$gw1,
                            gw2 = params[i,]$gw2,
                            pa = params[i,]$pa,
                            po = params[i,]$po,
                            vgseng1 = params[i,]$vgseng1,
                            vgseng2 = params[i,]$vgseng2,
                            vgseng3 = params[i,]$vgseng3)
    
    run_rhessys_single(input_rhessys = input_rhessys,
                       hdr_files = input_hdr,
                       tec_data = input_tec_data,
                       cmd_pars = stdpars,
                       runID = i)
  } 
  
  end_time = Sys.time()
  end_time - start_time
```

Prepare Static, SSURGO, and DSM Inputs

```{r Static SSURGO DSM }
setwd(system.file("extdata/", package = "RHESSysIOinR"))
# prepare RHESSys for run


params <- read.table(file = "out/cwparamtable.txt")

params<- as.data.frame(params)

n <- nrow(params)

soildefs <- list.files("defs/", pattern = "soil_", full.names = TRUE)

input_rhessys_static = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32static.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32static",
  flowtable = "CWWS32static.flow",
  start = "2013 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32static",
  commandline_options = c())

input_tec_data = IOin_tec_std(start = "2015 11 1 1",
                              end = "2018 11 1 1",
                              output_state = FALSE)

input_hdr_static = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = c("defs/lu_undev.def","defs/lu_agriculture.def","defs/lu_raingarden.def","defs/lu_urban.def"),
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")


input_rhessys_ssurgo = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32ssurgo.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32ssurgo",
  flowtable = "CWWS32ssurgo.flow",
  start = "2013 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32ssurgo",
  commandline_options = c())

input_hdr_ssurgo = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = c("defs/lu_undev.def","defs/lu_agriculture.def","defs/lu_raingarden.def","defs/lu_urban.def"),
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")


input_rhessys_dsm = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32dsm.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32dsm",
  flowtable = "CWWS32dsm.flow",
  start = "2013 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32dsm",
  commandline_options = c())

input_hdr_dsm = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = c("defs/lu_undev.def","defs/lu_agriculture.def","defs/lu_raingarden.def","defs/lu_urban.def"),
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")
  
```

Run RHESSys Calibration with Static Maps

```{r Run RHESSYS in Parallel using foreach Static}
setwd(system.file("extdata/", package = "RHESSysIOinR"))
n.cores <- parallel::detectCores() - 4  
  
my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
doParallel::registerDoParallel(cl = my.cluster)

start_time = Sys.time()
  
  foreach(i = 1:n, .packages = 'RHESSysIOinR') %dopar% {
    
    stdpars<- IOin_cmd_pars(m = params[i,]$m,
                            k = params[i,]$k,
                            soil_dep= params[i,]$soil_dep,
                            m_v = params[i,]$m_v,
                            k_v = params[i,]$k_v,
                            gw1 = params[i,]$gw1,
                            gw2 = params[i,]$gw2,
                            pa = params[i,]$pa,
                            po = params[i,]$po,
                            vgseng1 = params[i,]$vgseng1,
                            vgseng2 = params[i,]$vgseng2,
                            vgseng3 = params[i,]$vgseng3)
    
    run_rhessys_single(input_rhessys = input_rhessys_static,
                       hdr_files = input_hdr_static,
                       tec_data = input_tec_data,
                       cmd_pars = stdpars,
                       output_filter = basinfilter_static,
                       runID = i)
  } 
  
  end_time = Sys.time()
  end_time - start_time
```

Run RHESSys with SSURGO Maps

```{r Run RHESSYS in Parallel using foreach SSURGO}
setwd(system.file("extdata/", package = "RHESSysIOinR"))
n.cores <- parallel::detectCores() - 4  
  
my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
doParallel::registerDoParallel(cl = my.cluster)

start_time = Sys.time()
  
  foreach(i = 1:n, .packages = 'RHESSysIOinR') %dopar% {
    
    stdpars<- IOin_cmd_pars(m = params[i,]$m,
                            k = params[i,]$k,
                            soil_dep= params[i,]$soil_dep,
                            m_v = params[i,]$m_v,
                            k_v = params[i,]$k_v,
                            gw1 = params[i,]$gw1,
                            gw2 = params[i,]$gw2,
                            pa = params[i,]$pa,
                            po = params[i,]$po,
                            vgseng1 = params[i,]$vgseng1,
                            vgseng2 = params[i,]$vgseng2,
                            vgseng3 = params[i,]$vgseng3)
    
    run_rhessys_single(input_rhessys = input_rhessys_ssurgo,
                       hdr_files = input_hdr_ssurgo,
                       tec_data = input_tec_data,
                       cmd_pars = stdpars,
                       output_filter = basinfilter_ssurgo,
                       runID = i)
  } 
  
  end_time = Sys.time()
  end_time - start_time
```

Run RHESSys with RSS Maps

```{r Run RHESSYS in Parallel using foreach DSM}

setwd(system.file("extdata/", package = "RHESSysIOinR"))
n.cores <- parallel::detectCores() - 4  
  
my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
doParallel::registerDoParallel(cl = my.cluster)

start_time = Sys.time()
  
  foreach(i = 1:n, .packages = 'RHESSysIOinR') %dopar% {
    
    stdpars<- IOin_cmd_pars(m = params[i,]$m,
                            k = params[i,]$k,
                            soil_dep= params[i,]$soil_dep,
                            m_v = params[i,]$m_v,
                            k_v = params[i,]$k_v,
                            gw1 = params[i,]$gw1,
                            gw2 = params[i,]$gw2,
                            pa = params[i,]$pa,
                            po = params[i,]$po,
                            vgseng1 = params[i,]$vgseng1,
                            vgseng2 = params[i,]$vgseng2,
                            vgseng3 = params[i,]$vgseng3)
    
    run_rhessys_single(input_rhessys = input_rhessys_dsm,
                       hdr_files = input_hdr_dsm,
                       tec_data = input_tec_data,
                       cmd_pars = stdpars,
                       output_filter = basinfilter_dsm,
                       runID = i)
  } 
  
  end_time = Sys.time()
  end_time - start_time
```

Read in RHESSys Calibration Outputs and Plot

Calibration is performed using GLUE methodology and Nash-Sutcliffe Model Efficiency (NSE), Log-Normal Nash-Sutcliffe Model Efficiency (lnNSE), and Kling-Gupta Efficiency (KGE).

```{r Read in RHESSys Calibration Outputs and Plot, fig.height=9, fig.width=14, paged.print=TRUE}
 
setwd(system.file("extdata/", package = "RHESSysIOinR"))

# Read in RHESSys Calibration runs
#this is only reading in 1 patch

for(i in 1:n) { static_df <- read.csv(paste0("outputfilter/static_basin_daily_",i,".csv"))
  ssurgo_df <- read.csv(paste0("outputfilter/ssurgo_basin_daily_",i,".csv"))
  dsm_df <- read.csv(paste0("outputfilter/dsm_basin_daily_",i,".csv"))
  
  static_df$date <- as.Date(paste(static_df$year, static_df$month, static_df$day, sep = "-"), format = "%Y-%m-%d")
  ssurgo_df$date <- as.Date(paste(ssurgo_df$year, ssurgo_df$month, ssurgo_df$day, sep = "-"), format = "%Y-%m-%d")
  dsm_df$date <- as.Date(paste(dsm_df$year, dsm_df$month, dsm_df$day, sep = "-"), format = "%Y-%m-%d")
  
  static_df$patchstreamflow<- static_df$streamflow
  static_df$streamflow<- (static_df$patchstreamflow + static_df$gw.Qout)*1000
  
  ssurgo_df$patchstreamflow<- ssurgo_df$streamflow
  ssurgo_df$streamflow<- (ssurgo_df$patchstreamflow + ssurgo_df$gw.Qout)*1000
  
  dsm_df$patchstreamflow<- dsm_df$streamflow
  dsm_df$streamflow<- (dsm_df$patchstreamflow + dsm_df$gw.Qout)*1000
  
  assign(paste0("cwws32_static_run",i), static_df)
  assign(paste0("cwws32_ssurgo_run",i), ssurgo_df)
  assign(paste0("cwws32_dsm_run",i), dsm_df) }

#Prepare observed Streamflow data

obsws32<- read.csv(paste0(filedir,"input_files/streamflow/Daily Streamflow Watershed 32/Daily Streamflow Watershed 32/ObservedWS32.csv"))

obsws32$Date<- as.Date(obsws32$Date)

Caldates<- read.csv("outputfilter/caldates.csv")
Valdates<- read.csv("outputfilter/valdates.csv")

Caldates<- Caldates[,2]
Valdates<- Valdates[,2]

mergedsm<- read.table("outputfilter/mergedsm.csv", header = T, sep = ",")
obsws32sm<- read.table("outputfilter/obsws32sm.csv", header = T, sep = ",")
obsws32sm$Date<- as.Date(obsws32sm$Date,"%Y-%m-%d")

## plot observed data for period of record and calibration/validation periods as defined by SM data
{plot(obsws32$Date,obsws32$discharge_mm, type = 'l', xlim=c(as.Date(input_rhessys_static$start_date, "%Y%m%d"),as.Date(input_rhessys_static$end_date, "%Y%m%d")),ylab = "Observed Discharge mm/d", xlab = "Date", main = "Observed Streamflow Discharge for period of time model is run")
abline(v=as.Date(Caldates[1]), col = 'red', lwd = 2, lty = 2)
abline(v=as.Date(Caldates[2]), col = 'red', lwd = 2, lty = 2)
abline(v=as.Date(Valdates[1]), col = 'blue', lwd = 2, lty = 5)
abline(v=as.Date(Valdates[2]), col = 'blue', lwd = 2, lty = 5)
legend("topright",inset = 0.05, legend=c("Calibration","Validation"), col=c("red","blue"), lty=1)}

```

Calculate calibration performance metrics for STATIC

using 

rootzone.S/rootzone.field_capacity for soil moisture vwc


#must change to *

```{r calculate fit values and append them to a list, fig.height=9, fig.width=14, paged.print=TRUE}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

#creates cal subset, val subset, and merged streamflow

for(i in 1:n) { assign(paste0("simmerge",i), merge(obsws32,eval(parse(text = paste0("cwws32_static_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("simmergesm",i), merge(obsws32sm,eval(parse(text = paste0("cwws32_static_run",i))), by.x = "Date",by.y="date", all = FALSE))
  assign(paste0("calsubset",i),eval(parse(text = paste0("simmerge",i)))[eval(parse(text = paste0("simmerge",i,"$Date"))) >= Caldates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Caldates[2], ])
  assign(paste0("valsubset",i),eval(parse(text = paste0("simmerge",i)))[eval(parse(text = paste0("simmerge",i,"$Date"))) >= Valdates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Valdates[2], ])
  assign(paste0("calsubsetsm",i),eval(parse(text = paste0("simmergesm",i)))[eval(parse(text = paste0("simmergesm",i,"$Date"))) >= Caldates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Caldates[2], ])
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("simmergesm",i)))[eval(parse(text = paste0("simmergesm",i,"$Date"))) >= Valdates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Valdates[2], ])}

## create lists

NSElist <- lnNSElist <- KGElist <- smNSElist <- smlnNSElist <- smKGElist <- list()

for(i in 1:n) { assign(paste0("NSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow")))) ))
  NSElist[[i]]<- eval(parse(text = paste0("NSEobs",i))) }

for(i in 1:n) {  assign(paste0("lnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  lnNSElist[[i]]<- eval(parse(text = paste0("lnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("KGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow"))))))
  KGElist[[i]]<- eval(parse(text = paste0("KGEobs",i))) }

for(i in 1:n) {  assign(paste0("smNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture"))))))
  smNSElist[[i]]<- eval(parse(text = paste0("smNSEobs",i))) }

for(i in 1:n) {  assign(paste0("smlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  smlnNSElist[[i]]<- eval(parse(text = paste0("smlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("smKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture"))))))
  smKGElist[[i]]<- eval(parse(text = paste0("smKGEobs",i))) }

## modified for calsubset

calNSElist<- callnNSElist <- calKGElist <- calsmNSElist<- calsmlnNSElist<- calsmKGElist<- list()

for(i in 1:n) { assign(paste0("cNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow")))) ))
  calNSElist[[i]]<- eval(parse(text = paste0("cNSEobs",i))) }

for(i in 1:n) {  assign(paste0("clnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  callnNSElist[[i]]<- eval(parse(text = paste0("clnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("cKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow"))))))
  calKGElist[[i]]<- eval(parse(text = paste0("cKGEobs",i))) }

for(i in 1:n) {  assign(paste0("csmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture"))))))
  calsmNSElist[[i]]<- eval(parse(text = paste0("csmNSEobs",i))) }

for(i in 1:n) {  assign(paste0("csmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  calsmlnNSElist[[i]]<- eval(parse(text = paste0("csmlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("csmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture"))))))
  calsmKGElist[[i]]<- eval(parse(text = paste0("csmKGEobs",i))) }

## modified for val subset

valNSElist<- vallnNSElist<- valKGElist<- valsmNSElist <- valsmlnNSElist<- valsmKGElist <- list()

for(i in 1:n) { assign(paste0("vNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow")))) ))
  valNSElist[[i]]<- eval(parse(text = paste0("vNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vlnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  vallnNSElist[[i]]<- eval(parse(text = paste0("vlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow"))))))
  valKGElist[[i]]<- eval(parse(text = paste0("vKGEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
  valsmKGElist[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }

Reduce(max, NSElist)
Reduce(max, lnNSElist)
Reduce(max, KGElist)

Reduce(max, smNSElist)
Reduce(max, smlnNSElist)
Reduce(max, smKGElist)

which(NSElist == Reduce(max, NSElist), arr.ind=TRUE)
which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE)
which(KGElist == Reduce(max, KGElist), arr.ind=TRUE)

which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE)
which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE)
which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE)

paramtable<- read.table(file = "out/cwparamtable.txt")

paramtable[which(NSElist == Reduce(max, NSElist), arr.ind=TRUE),]
paramtable[which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE),]
paramtable[which(KGElist == Reduce(max, KGElist), arr.ind=TRUE),]

paramtable[which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE),]
paramtable[which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE),]
paramtable[which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE),]

###plot all data for GLUE

paramtable$run<- seq.int(nrow(paramtable))
paramtable$rowname<- row.names(paramtable)

paramtable$NSE<- NSElist
paramtable$lnNSE<- lnNSElist
paramtable$KGE<- KGElist

paramtable$smNSE<- smNSElist
paramtable$smlnNSE<- smlnNSElist
paramtable$smKGE<- smKGElist

#add in calibration values

paramtable$calNSE<- calNSElist
paramtable$callnNSE<- callnNSElist
paramtable$calKGE<- calKGElist

paramtable$calsmNSE<- calsmNSElist
paramtable$calsmlnNSE<- calsmlnNSElist
paramtable$calsmKGE<- calsmKGElist

#add in validation values
paramtable$valNSE<- valNSElist
paramtable$vallnNSE<- vallnNSElist
paramtable$valKGE<- valKGElist

paramtable$valsmNSE<- valsmNSElist
paramtable$valsmlnNSE<- valsmlnNSElist
paramtable$valsmKGE<- valsmKGElist

paramtable<- as.data.frame(lapply(paramtable,unlist))


plot(paramtable$KGE)
plot(paramtable$smKGE)


write.table(paramtable, file = "out/staticcalibrationoutput.txt")
```
Calculate calibration performance metrics for SSURGO

using 

rootzone.S/rootzone.field_capacity for soil moisture vwc

```{r calculate fit values and append them to a list ssurgo, fig.height=9, fig.width=14, paged.print=TRUE}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

##merge calibration runs with observed so dates match up, instead of subsetting 

#creates cal subset, val subset, and merged streamflow

for(i in 1:n) { assign(paste0("simmerge",i), merge(obsws32,eval(parse(text = paste0("cwws32_ssurgo_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("simmergesm",i), merge(obsws32sm,eval(parse(text = paste0("cwws32_ssurgo_run",i))), by.x = "Date",by.y="date", all = FALSE))
  assign(paste0("calsubset",i),eval(parse(text = paste0("simmerge",i)))[eval(parse(text = paste0("simmerge",i,"$Date"))) >= Caldates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Caldates[2], ])
  assign(paste0("valsubset",i),eval(parse(text = paste0("simmerge",i)))[eval(parse(text = paste0("simmerge",i,"$Date"))) >= Valdates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Valdates[2], ])
  assign(paste0("calsubsetsm",i),eval(parse(text = paste0("simmergesm",i)))[eval(parse(text = paste0("simmergesm",i,"$Date"))) >= Caldates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Caldates[2], ])
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("simmergesm",i)))[eval(parse(text = paste0("simmergesm",i,"$Date"))) >= Valdates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Valdates[2], ])}

## create lists


NSElist <- lnNSElist <- KGElist <- smNSElist <- smlnNSElist <- smKGElist <- list()

for(i in 1:n) { assign(paste0("NSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow")))) ))
  NSElist[[i]]<- eval(parse(text = paste0("NSEobs",i))) }

for(i in 1:n) {  assign(paste0("lnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  lnNSElist[[i]]<- eval(parse(text = paste0("lnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("KGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow"))))))
  KGElist[[i]]<- eval(parse(text = paste0("KGEobs",i))) }

for(i in 1:n) {  assign(paste0("smNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture"))))))
  smNSElist[[i]]<- eval(parse(text = paste0("smNSEobs",i))) }

for(i in 1:n) {  assign(paste0("smlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  smlnNSElist[[i]]<- eval(parse(text = paste0("smlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("smKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture"))))))
  smKGElist[[i]]<- eval(parse(text = paste0("smKGEobs",i))) }

## modified for calsubset

calNSElist<- callnNSElist <- calKGElist <- calsmNSElist<- calsmlnNSElist<- calsmKGElist<- list()

for(i in 1:n) { assign(paste0("cNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow")))) ))
  calNSElist[[i]]<- eval(parse(text = paste0("cNSEobs",i))) }

for(i in 1:n) {  assign(paste0("clnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  callnNSElist[[i]]<- eval(parse(text = paste0("clnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("cKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow"))))))
  calKGElist[[i]]<- eval(parse(text = paste0("cKGEobs",i))) }

for(i in 1:n) {  assign(paste0("csmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture"))))))
  calsmNSElist[[i]]<- eval(parse(text = paste0("csmNSEobs",i))) }

for(i in 1:n) {  assign(paste0("csmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  calsmlnNSElist[[i]]<- eval(parse(text = paste0("csmlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("csmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture"))))))
  calsmKGElist[[i]]<- eval(parse(text = paste0("csmKGEobs",i))) }

## modified for val subset

valNSElist<- vallnNSElist<- valKGElist<- valsmNSElist <- valsmlnNSElist<- valsmKGElist <- list()

for(i in 1:n) { assign(paste0("vNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow")))) ))
  valNSElist[[i]]<- eval(parse(text = paste0("vNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vlnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  vallnNSElist[[i]]<- eval(parse(text = paste0("vlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow"))))))
  valKGElist[[i]]<- eval(parse(text = paste0("vKGEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
  valsmKGElist[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }

Reduce(max, NSElist)
Reduce(max, lnNSElist)
Reduce(max, KGElist)

Reduce(max, smNSElist)
Reduce(max, smlnNSElist)
Reduce(max, smKGElist)

which(NSElist == Reduce(max, NSElist), arr.ind=TRUE)
which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE)
which(KGElist == Reduce(max, KGElist), arr.ind=TRUE)

which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE)
which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE)
which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE)

paramtable<- read.table(file = "out/cwparamtable.txt")

paramtable[which(NSElist == Reduce(max, NSElist), arr.ind=TRUE),]
paramtable[which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE),]
paramtable[which(KGElist == Reduce(max, KGElist), arr.ind=TRUE),]

paramtable[which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE),]
paramtable[which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE),]
paramtable[which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE),]

###plot all data for GLUE

paramtable$run<- seq.int(nrow(paramtable))
paramtable$rowname<- row.names(paramtable)

paramtable$NSE<- NSElist
paramtable$lnNSE<- lnNSElist
paramtable$KGE<- KGElist

paramtable$smNSE<- smNSElist
paramtable$smlnNSE<- smlnNSElist
paramtable$smKGE<- smKGElist

#add in calibration values

paramtable$calNSE<- calNSElist
paramtable$callnNSE<- callnNSElist
paramtable$calKGE<- calKGElist

paramtable$calsmNSE<- calsmNSElist
paramtable$calsmlnNSE<- calsmlnNSElist
paramtable$calsmKGE<- calsmKGElist

#add in validation values
paramtable$valNSE<- valNSElist
paramtable$vallnNSE<- vallnNSElist
paramtable$valKGE<- valKGElist

paramtable$valsmNSE<- valsmNSElist
paramtable$valsmlnNSE<- valsmlnNSElist
paramtable$valsmKGE<- valsmKGElist

paramtable<- as.data.frame(lapply(paramtable,unlist))

write.table(paramtable, file = "out/ssurgocalibrationoutput.txt")
```

Calculate calibration performance metrics for DSM

using 

rootzone.S/rootzone.field_capacity for soil moisture vwc

```{r calculate fit values and append them to a list dsm, fig.height=9, fig.width=14, paged.print=TRUE}
setwd(system.file("extdata/", package = "RHESSysIOinR"))


##merge calibration runs with observed so dates match up, instead of subsetting 

#creates cal subset, val subset, and merged streamflow

for(i in 1:n) { assign(paste0("simmerge",i), merge(obsws32,eval(parse(text = paste0("cwws32_dsm_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("simmergesm",i), merge(obsws32sm,eval(parse(text = paste0("cwws32_dsm_run",i))), by.x = "Date",by.y="date", all = FALSE))
  assign(paste0("calsubset",i),eval(parse(text = paste0("simmerge",i)))[eval(parse(text = paste0("simmerge",i,"$Date"))) >= Caldates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Caldates[2], ])
  assign(paste0("valsubset",i),eval(parse(text = paste0("simmerge",i)))[eval(parse(text = paste0("simmerge",i,"$Date"))) >= Valdates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Valdates[2], ])
  assign(paste0("calsubsetsm",i),eval(parse(text = paste0("simmergesm",i)))[eval(parse(text = paste0("simmergesm",i,"$Date"))) >= Caldates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Caldates[2], ])
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("simmergesm",i)))[eval(parse(text = paste0("simmergesm",i,"$Date"))) >= Valdates[1] & eval(parse(text = paste0("simmerge",i,"$Date"))) <= Valdates[2], ])}

## create lists


NSElist <- lnNSElist <- KGElist <- smNSElist <- smlnNSElist <- smKGElist <- list()

for(i in 1:n) { assign(paste0("NSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow")))) ))
  NSElist[[i]]<- eval(parse(text = paste0("NSEobs",i))) }

for(i in 1:n) {  assign(paste0("lnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  lnNSElist[[i]]<- eval(parse(text = paste0("lnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("KGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("simmerge",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("simmerge",i,"$observedstreamflow"))))))
  KGElist[[i]]<- eval(parse(text = paste0("KGEobs",i))) }

for(i in 1:n) {  assign(paste0("smNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture"))))))
  smNSElist[[i]]<- eval(parse(text = paste0("smNSEobs",i))) }

for(i in 1:n) {  assign(paste0("smlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  smlnNSElist[[i]]<- eval(parse(text = paste0("smlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("smKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("simmergesm",i,"$rootzone.S","*simmergesm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("simmergesm",i,"$mergedsoilmoisture"))))))
  smKGElist[[i]]<- eval(parse(text = paste0("smKGEobs",i))) }

## modified for calsubset

calNSElist<- callnNSElist <- calKGElist <- calsmNSElist<- calsmlnNSElist<- calsmKGElist<- list()

for(i in 1:n) { assign(paste0("cNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow")))) ))
  calNSElist[[i]]<- eval(parse(text = paste0("cNSEobs",i))) }

for(i in 1:n) {  assign(paste0("clnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  callnNSElist[[i]]<- eval(parse(text = paste0("clnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("cKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("calsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("calsubset",i,"$observedstreamflow"))))))
  calKGElist[[i]]<- eval(parse(text = paste0("cKGEobs",i))) }

for(i in 1:n) {  assign(paste0("csmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture"))))))
  calsmNSElist[[i]]<- eval(parse(text = paste0("csmNSEobs",i))) }

for(i in 1:n) {  assign(paste0("csmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  calsmlnNSElist[[i]]<- eval(parse(text = paste0("csmlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("csmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$rootzone.S","*calsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("calsubsetsm",i,"$mergedsoilmoisture"))))))
  calsmKGElist[[i]]<- eval(parse(text = paste0("csmKGEobs",i))) }

## modified for val subset

valNSElist<- vallnNSElist<- valKGElist<- valsmNSElist <- valsmlnNSElist<- valsmKGElist <- list()

for(i in 1:n) { assign(paste0("vNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow")))) ))
  valNSElist[[i]]<- eval(parse(text = paste0("vNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vlnNSEobs",i), NSE( sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  vallnNSElist[[i]]<- eval(parse(text = paste0("vlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valsubset",i,"$streamflow")))), obs = as.numeric(eval(parse(text = paste0("valsubset",i,"$observedstreamflow"))))))
  valKGElist[[i]]<- eval(parse(text = paste0("vKGEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:n) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rootzone.S","*valsubsetsm",i,"$rootzone.field_capacity")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
  valsmKGElist[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }

Reduce(max, NSElist)
Reduce(max, lnNSElist)
Reduce(max, KGElist)

Reduce(max, smNSElist)
Reduce(max, smlnNSElist)
Reduce(max, smKGElist)

which(NSElist == Reduce(max, NSElist), arr.ind=TRUE)
which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE)
which(KGElist == Reduce(max, KGElist), arr.ind=TRUE)

which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE)
which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE)
which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE)

paramtable<- read.table(file = "out/cwparamtable.txt")

paramtable[which(NSElist == Reduce(max, NSElist), arr.ind=TRUE),]
paramtable[which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE),]
paramtable[which(KGElist == Reduce(max, KGElist), arr.ind=TRUE),]

paramtable[which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE),]
paramtable[which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE),]
paramtable[which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE),]

###plot all data for GLUE

paramtable$run<- seq.int(nrow(paramtable))
paramtable$rowname<- row.names(paramtable)

paramtable$NSE<- NSElist
paramtable$lnNSE<- lnNSElist
paramtable$KGE<- KGElist

paramtable$smNSE<- smNSElist
paramtable$smlnNSE<- smlnNSElist
paramtable$smKGE<- smKGElist

#add in calibration values

paramtable$calNSE<- calNSElist
paramtable$callnNSE<- callnNSElist
paramtable$calKGE<- calKGElist

paramtable$calsmNSE<- calsmNSElist
paramtable$calsmlnNSE<- calsmlnNSElist
paramtable$calsmKGE<- calsmKGElist

#add in validation values
paramtable$valNSE<- valNSElist
paramtable$vallnNSE<- vallnNSElist
paramtable$valKGE<- valKGElist

paramtable$valsmNSE<- valsmNSElist
paramtable$valsmlnNSE<- valsmlnNSElist
paramtable$valsmKGE<- valsmKGElist

paramtable<- as.data.frame(lapply(paramtable,unlist))

write.table(paramtable, file = "out/dsmcalibrationoutput.txt")
```

```{r create plots, fig.height=9, fig.width=14, paged.print=TRUE}
top100<- paramtable[order(-paramtable$lnNSE),][1:n,]

top100$run

### use values from top 100 run to table of daily streamflow predictions, for each day predict 95% prediction interval and mean then plot
#subset streamflow value for each day - day 1 day 2 etc
#put all values into a table

top100[["run"]]

toppredictions<- data.frame(cwws32_dsm_run1$date)

for (i in top100$run){
 # print(i)
  toppredictions[,ncol(toppredictions)+1] <- eval(parse(text=paste0("cwws32_dsm_run",i,"$streamflow")))
  colnames(toppredictions)[ncol(toppredictions)] <- paste0("stream",i)
}

head(toppredictions[1:5,])

toppredictions<- toppredictions[,-1]

toppredictions$upper<- 0
toppredictions$mean<- 0
toppredictions$lower<- 0

for (i in 1:nrow(toppredictions)){
 #  print(i)
  upper= CI(as.numeric(toppredictions[i,]),ci=0.95)[1]
  mean= CI(as.numeric(toppredictions[i,]),ci=0.95)[2]
  lower= CI(as.numeric(toppredictions[i,]),ci=0.95)[3]
  
  toppredictions[i,]$upper<- upper
  toppredictions[i,]$mean<- mean
  toppredictions[i,]$lower<- lower
}

toppredictions<-cbind(date = cwws32_dsm_run1$date,toppredictions)

head(toppredictions[1:5,])

# toppredictions[1:500,]

ggplot(toppredictions,aes(date,mean,group=1))+
  geom_line(col = 'red')+
  geom_ribbon(aes(ymin=lower, ymax=upper), fill = "blue", alpha=0.3)+
  geom_line(data=calsubset1,aes(x=Date,y=discharge_mm))+
  ylab("Discharge mm")+
  xlab("Date")+
  labs(title = "Top 100 Model Runs vs Measured Streamflow")


par(mfrow=c(1, 1))

old.par <- par(mfrow=c(3, 7), mar = c(5,4,4,2))

{plot(paramtable$m,paramtable$NSE,xlab = "m parameter", pch = 20)
lines(lowess(paramtable$m,paramtable$NSE, f = 0.1),col="red",ylim = c(-1,0.7))
plot(paramtable$k,paramtable$NSE,xlab = "lateral ksat parameter", pch = 20)
lines(lowess(paramtable$k,paramtable$NSE, f = 0.1),col="red",ylim = c(-1,0.7))
plot(paramtable$soil_dep,paramtable$NSE,xlab = "sdepth parameter", pch = 20)
lines(lowess(paramtable$soil_dep,paramtable$NSE, f = 0.1),col="red",ylim = c(-1,0.7))
plot(paramtable$m_v,paramtable$NSE,xlab = "mv parameter", pch = 20)
lines(lowess(paramtable$m_v,paramtable$NSE, f = 0.1),col="red",ylim = c(-1,0.7))
plot(paramtable$k_v,paramtable$NSE,xlab = "ksatv parameter", pch = 20)
lines(lowess(paramtable$k_v,paramtable$NSE, f = 0.1),col="red",ylim = c(-1,0.7))
plot(paramtable$gw1,paramtable$NSE,xlab = "gw1 parameter", pch = 20)
lines(lowess(paramtable$gw1,paramtable$NSE, f = 0.1),col="red",ylim = c(-1,0.7))
plot(paramtable$gw2,paramtable$NSE,xlab = "gw2 parameter", pch = 20)
lines(lowess(paramtable$gw2,paramtable$NSE, f = 0.1),col="red",ylim = c(-1,0.7))}


{plot(paramtable$m,paramtable$lnNSE,ylim = c(-1,0.7),xlab = "m parameter", pch = 20)
lines(lowess(paramtable$m,paramtable$lnNSE, f = 0.1),col="red")
plot(paramtable$k,paramtable$lnNSE,ylim = c(-1,0.7),xlab = "lateral ksat parameter", pch = 20)
lines(lowess(paramtable$k,paramtable$lnNSE, f = 0.1),col="red")
plot(paramtable$soil_dep,paramtable$lnNSE,ylim = c(-1,0.7),xlab = "sdepth parameter", pch = 20)
lines(lowess(paramtable$soil_dep,paramtable$lnNSE, f = 0.1),col="red")
plot(paramtable$m_v,paramtable$lnNSE,ylim = c(-1,0.7),xlab = "mv parameter", pch = 20)
lines(lowess(paramtable$m_v,paramtable$lnNSE, f = 0.1),col="red")
plot(paramtable$k_v,paramtable$lnNSE,ylim = c(-1,0.7),xlab = "ksatv parameter", pch = 20)
lines(lowess(paramtable$k_v,paramtable$lnNSE, f = 0.1),col="red")
plot(paramtable$gw1,paramtable$lnNSE,ylim = c(-1,0.7),xlab = "gw1 parameter", pch = 20)
lines(lowess(paramtable$gw1,paramtable$lnNSE, f = 0.1),col="red")
plot(paramtable$gw2,paramtable$lnNSE,ylim = c(-1,0.7),xlab = "gw2 parameter", pch = 20)
lines(lowess(paramtable$gw2,paramtable$lnNSE, f = 0.1),col="red")}


{plot(paramtable$m,paramtable$KGE,ylim = c(-0.25,1),xlab = "m parameter", pch = 20)
lines(lowess(paramtable$m,paramtable$KGE, f = 0.1),col="red")
plot(paramtable$k,paramtable$KGE,ylim = c(-0.25,1),xlab = "lateral ksat parameter", pch = 20)
lines(lowess(paramtable$k,paramtable$KGE, f = 0.1),col="red")
plot(paramtable$soil_dep,paramtable$KGE,ylim = c(-0.25,1),xlab = "sdepth parameter", pch = 20)
lines(lowess(paramtable$soil_dep,paramtable$KGE, f = 0.1),col="red")
plot(paramtable$m_v,paramtable$KGE,ylim = c(-0.25,0.7),xlab = "mv parameter", pch = 20)
lines(lowess(paramtable$m_v,paramtable$KGE, f = 0.1),col="red")
plot(paramtable$k_v,paramtable$KGE,ylim = c(-0.25,0.7),xlab = "ksatv parameter", pch = 20)
lines(lowess(paramtable$k_v,paramtable$KGE, f = 0.1),col="red")
plot(paramtable$gw1,paramtable$KGE,ylim = c(-0.25,0.7),xlab = "gw1 parameter", pch = 20)
lines(lowess(paramtable$gw1,paramtable$KGE, f = 0.1),col="red")
plot(paramtable$gw2,paramtable$KGE,ylim = c(-0.25,0.7),xlab = "gw2 parameter", pch = 20)
lines(lowess(paramtable$gw2,paramtable$KGE, f = 0.1),col="red")}

par(old.par)

old.par <- par(mfrow=c(1, 12), mar = c(5,4,4,2))

{plot(paramtable$m,paramtable$smKGE,ylim = c(-0.25,1),xlab = "m parameter", pch = 20)
lines(lowess(paramtable$m,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$k,paramtable$smKGE,ylim = c(-0.25,1),xlab = "lateral ksat parameter", pch = 20)
lines(lowess(paramtable$k,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$soil_dep,paramtable$smKGE,ylim = c(-0.25,1),xlab = "sdepth parameter", pch = 20)
lines(lowess(paramtable$soil_dep,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$m_v,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "mv parameter", pch = 20)
lines(lowess(paramtable$m_v,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$k_v,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "ksatv parameter", pch = 20)
lines(lowess(paramtable$k_v,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$gw1,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "gw1 parameter", pch = 20)
lines(lowess(paramtable$gw1,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$gw2,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "gw2 parameter", pch = 20)
lines(lowess(paramtable$gw2,paramtable$smKGE, f = 0.1),col="red")}

{plot(paramtable$pa,paramtable$smKGE,ylim = c(-0.25,1),xlab = "pore size index", pch = 20)
lines(lowess(paramtable$pa,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$po,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "psi air entry", pch = 20)
lines(lowess(paramtable$po,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$vgseng1,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "vgseng1", pch = 20)
lines(lowess(paramtable$vgseng1,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$vgseng2,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "vgseng2", pch = 20)
lines(lowess(paramtable$vgseng2,paramtable$smKGE, f = 0.1),col="red")
plot(paramtable$vgseng3,paramtable$smKGE,ylim = c(-0.25,0.7),xlab = "vgseng3", pch = 20)
lines(lowess(paramtable$vgseng3,paramtable$smKGE, f = 0.1),col="red")}

par(old.par)

### end plot all data for GLUE

paste0("cwws32_dsm_run",which(NSElist == Reduce(max, NSElist), arr.ind=TRUE),"$streamflow")
paste0("cwws32_dsm_run",which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE),"$streamflow")
paste0("cwws32_dsm_run",which(KGElist == Reduce(max, KGElist), arr.ind=TRUE),"$streamflow")


{plot(cwws32_dsm_run3$date,cwws32_dsm_run3$streamflow, type = "l", col = 'red', main = "Observed Streamflow vs Top Runs", xlab = "Date", ylab = "Streamflow")
lines(cwws32_dsm_run5$date,cwws32_dsm_run5$streamflow, col = 'blue')
#lines(cwws32_dsm_run46$bd$date,cwws32_dsm_run46$bd$streamflow, col = 'green')
lines(simmerge1$Date,simmerge1$observedstreamflow, col = 'orange', lwd = 2, lty = 2)}


#{plot(paste("cwws32_run",which(NSElist == Reduce(max, NSElist), arr.ind=TRUE),"$bd$date", sep = ""),paste("cwws32_run",which(NSElist == Reduce(max, NSElist), arr.ind=TRUE),"$bd$streamflow", sep = ""), type = "l", col = 'red', main = "Observed Streamflow vs Top Runs", xlab = "Date", ylab = "Streamflow")
#lines(paste("cwws32_run",which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE),"$bd$date", sep = ""),paste("cwws32_run",which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE),"$bd$streamflow", sep = ""), col = 'blue')
#lines(cwws32_run80$bd$date,cwws32_run80$bd$streamflow, col = 'green')
#lines(paste("cwws32_run",which(KGElist == Reduce(max, KGElist), arr.ind=TRUE),"$bd$date", sep = ""),paste("cwws32_run",which(KGElist == Reduce(max, KGElist), arr.ind=TRUE),"$bd$streamflow", sep = ""), col = 'orange', lwd = 2, lty = 2)
#legend("topleft",inset = 0.05, legend=c("Oberved Discharge","Best NSE","BEST lnNSE","Best KGE"), col=c("orange","red","blue","green"), lty=c(2,1,1,1))}

{plot(cwws32_dsm_run16$date,cwws32_dsm_run16$streamflow, type = "l", col = 'red', main = "Observed Streamflow vs Top Runs", xlab = "Date", ylab = "Streamflow", ylim = c(0,10))
#lines(cwws32_dsm_run47$bd$date,cwws32_dsm_run47$bd$streamflow, col = 'blue')
lines(obsws32$Date,obsws32$discharge_mm, col = 'black', lwd = 2, lty = 2)}

```


```{r}

which(NSElist == Reduce(max, NSElist), arr.ind=TRUE)
which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE)
which(KGElist == Reduce(max, KGElist), arr.ind=TRUE)

which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE)
which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE)
which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE)

paramtable[which(NSElist == Reduce(max, NSElist), arr.ind=TRUE),]
paramtable[which(lnNSElist == Reduce(max, lnNSElist), arr.ind=TRUE),]
paramtable[which(KGElist == Reduce(max, KGElist), arr.ind=TRUE),]

paramtable[which(smNSElist == Reduce(max, smNSElist), arr.ind=TRUE),]
paramtable[which(smlnNSElist == Reduce(max, smlnNSElist), arr.ind=TRUE),]
paramtable[which(smKGElist == Reduce(max, smKGElist), arr.ind=TRUE),]
```
