==> Step5_Validation.Rmd <==
---
title: "RHESSysIOinR_Coweeta_WS32_74 Multi Site Validation Field Data_Step5"
author: "Carlos Quintero"
date: "6/17/2024"
output:
  html_document: default
  pdf_document: default
---
```{r include=FALSE}
knitr::opts_knit$set(root.dir = system.file("extdata/", package = "RHESSysIOinR"))
 setwd(system.file("extdata/", package = "RHESSysIOinR"))
```

Coweeta Hydrologic Laboratory Watershed 32 using Static, SSURGO, and RSS soil inputs. 

The following R Markdown script prepares and runs the Regional Hydro-Ecologic Simulation System (RHESSYS) within the R environment. 
Modified from RHESSysIOinR example scripts created by Will Burke and Ryan Bart. Includes CLHS code snippets and SDA code from Dylan Beaudette.

Windows Subsystem via Linux must be installed before RHESSys will run in a Windows environment.
GRASS8 must be installed to run the spatial preprocessing portion of the script.

Code has been modified to run RHESSys 7.4 and 5.18. 
7.4 vegetation processing differs slightly from 5.18.
Various template differences exist between the two versions of RHESSys and template files must be carefully reviewed if moving between versions.
This script is currently set up to run RHESSys 7.4

```{r setup, include=TRUE, verbose = FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
#install.packages("devtools")
#devtools::install_github("ropensci/FedData")

library(aqp)
library(daymetr)
library(clhs)
library(EcoHydRology)
library(elevatr)
library(ezknitr)
#library(FedData)
library(foreach)
library(foreign)
library(ggplot2)
library(ggpubr)
library(hydroGOF)
library(imputeTS)
library(knitr)
library(latticeExtra)
library(lubridate)
library(Metrics)
library(raster)
library(rasterVis)
library(readxl)
library(reshape2)
library(rgdal)
library(rgrass)
library(rmarkdown)
library(Rmisc)
library(RHESSysPreprocessing)
library(RHESSysIOinR)
library(sf)
library(soilDB)
library(sp)
library(stats)
library(stringi)
library(tactile)
library(terra)
library(testthat)
library(tidyverse)
library(viridisLite)
library(dplyr)

setwd(system.file("extdata/", package = "RHESSysIOinR"))
getwd()
rh_ver = dir(path = "modelfiles/rhessys/", pattern = "^rhessys\\d+",recursive = F)
rh_path = file.path("modelfiles/rhessys", rh_ver)

## Load in necessary files
filedir <- ("C:/Users/Carlos/Documents/GitHub/RHESSys74-R-Markdown/")

watershedshapefile <- paste0(filedir, "input_files/shapefiles/watersheds/Coweeta_Hydrologic_Laboratory.shp")
streamshapefile<- paste0(filedir,"input_files/shapefiles/streams/StreamsCaldwellClipped.shp")
weirshapefile<- paste0(filedir,"input_files/shapefiles/weirs/Subbasinoutlets.shp")
roadshapefile <- paste0(filedir, "input_files/shapefiles/roads/UpdatedRoads4Reprojected32617.shp")

CWRG12 <- paste0(filedir,"input_files/precipitation/RDS-2017-0031/Data/RG12_daily_1942_2021.csv")
CWRG20 <-paste0(filedir,"input_files/precipitation/RDS-2017-0031/Data/RG20_daily_1962_2021.csv")

CS21<- paste0(filedir,"input_files/temperature/RDS-2015-0042/Data/cs21_daily.csv")
CS28<- paste0(filedir,"input_files/temperature/RDS-2015-0042/Data/cs28_daily.csv")

newsmlocations<-  paste0(filedir,"input_files/shapefiles/soil_moisture/CoweetaNRCSsmKML.kml")

```
#in step 4

```{r}

setwd(system.file("extdata/", package = "RHESSysIOinR"))
#Prepare observed Streamflow data


obsws32<- read.csv(paste0(filedir,"input_files/streamflow/Daily Streamflow Watershed 32/Daily Streamflow Watershed 32/ObservedWS32.csv"))


obsws32$Date<- as.Date(obsws32$Date)

Caldates<- read.csv("outputfilter/caldates.csv")
Valdates<- read.csv("outputfilter/valdates.csv")

Caldates<- Caldates[,2]
Valdates<- Valdates[,2]

mergedsm<- read.table("outputfilter/mergedsm.csv", header = T, sep = ",")
obsws32sm<- read.table("outputfilter/obsws32sm.csv", header = T, sep = ",")
obsws32sm$Date<- as.Date(obsws32sm$Date,"%Y-%m-%d")
```

New Patch Filter

Basin - Basin 
Hillslope - Subbasin
Zone - Patch
Patch - Patch 

1 112 136557
1 110 141942
1 110 149478

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

patchfilterpatch1 = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "valws32patch1",
                                   spatial_level = "patch",
                                   spatial_ID = "1:70:144625:144625",
                                   variables = c("rootzone.S","rootzone.potential_sat","rootzone.field_capacity", "rz_storage"))

patchfilterpatch2 = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "valws32patch2",
                                   spatial_level = "patch",
                                   spatial_ID = "1:70:149472:149472",
                                   variables = c("rootzone.S","rootzone.potential_sat","rootzone.field_capacity", "rz_storage"))

patchfilterpatch3 = build_output_filter(timestep = "daily",
                                   output_format = "csv",
                                   output_path = "outputfilter",
                                   output_filename = "valws32patch3",
                                   spatial_level = "patch",
                                   spatial_ID = "1:70:157009:157009",
                                   variables = c("rootzone.S","rootzone.potential_sat","rootzone.field_capacity", "rz_storage"))


# start_time = Sys.time()
# start_time
# run_rhessys_single(input_rhessys = input_rhessys,
#                    hdr_files = input_hdr,
#                    tec_data = input_tec_data,
#                    cmd_pars = stdpars,
#                    output_filter = c(patchfilterpatch1,patchfilterpatch2,patchfilterpatch3),
#                    runID = 'single')
# end_time = Sys.time()
# end_time - start_time
# end_time


```

*** VALIDATION ***

Run Model once for validation, copied from previous code must clean up


Configure RHESSys Inputs/Outputs for a single run.

Use World Statefile created during spin-up. To run model one time at patch scale.

 COMMAND LINE OPTIONS

-b		Basin output option.  Print out response variables for specified basins. 
-c		Canopy stratum output option.  Print out response variables for specified strata. 
-g		Grow option.  Try to read in dynamic bgc input data and output dynamic bgc parameters. 
-h    Hillslope output option.  Print out response variables for specified hillslopes. 
-p		Patch output option.  Print out response variables for specified patches. 
-r		Routing option. Gives name of flow_table to define explicit routing connectivity.  Also trigger use of explicit routing over TOPMODEL approach. 
-c		Stratum output option.  Print out response variables for specified strata. 

patches listed top to bottom

```{r Configure RHESSys Inputs for Validation}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

soildefs <- list.files("defs/", pattern = "soil_", full.names = TRUE)

input_rhessys = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32dsm.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32dsm",
  flowtable = "CWWS32dsm.flow",
  start = "2013 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32",
 commandline_options = c("-b -g -p"))

## TEC file dictates model output, begin output a year in to allow model SM to stabilize
# do not output_state or worldfile may be overwritten as output is created
input_tec_data = IOin_tec_std(start = "2015 11 1 1",
                              end = "2018 11 1 1",
                              output_state = FALSE)

input_hdr = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = "defs/lu_undev.def",
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")

## Calibrated Values 


stdpars<- IOin_cmd_pars(
            m = 1.581785,
            k = 41.12945,
            soil_dep= 0.1776117,
            m_v = 8.11574,
            k_v = 0.8713272,
            gw1 = 0.2477113,
            gw2 = 0.4291277,
            pa = 0.3913392,
            po = 1.85424,
            vgseng1 = 1.459184,
            vgseng2 = 1.799894,
            vgseng3 = 1.381696)


```

Run RHESSys once for validation

```{r Run Rhessys Once for Validation}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

start_time = Sys.time()
run_rhessys_single(input_rhessys = input_rhessys,
                   hdr_files = input_hdr,
                   tec_data = input_tec_data,
                   cmd_pars = stdpars,
                   runID = 'validation')
end_time = Sys.time()
end_time - start_time

```

Read RHESSys Single Run Output

```{r Read in RHESSys Validation Output}
setwd(system.file("extdata/", package = "RHESSysIOinR"))
getwd()

singlerunvalidation<- readin_rhessys_output("out/cwws32_runvalidation")

plot(singlerunvalidation$bd$streamflow~singlerunvalidation$bd$date, type = "l", main = "Streamflow", xlab = "Date", ylab = "Streamflow", col = 'DarkBlue')

plot(as.Date(singlerunvalidation$bd$date),singlerunvalidation$bd$rz_storage, type = "l", col = 'black', main = "Basin Scale Root Zone Storage",
    xlab = "Date", ylab = "mm")

#basin scale soil moisture
plot(singlerunvalidation$bd$rz_storage/singlerunvalidation$bdg$root_depth~singlerunvalidation$bd$date, type = "l", main = "RZ_Storage/Root_Depth x Date", xlab = "Date", ylab = "rz_Storage/root_depth", col = 'brown')

plot(singlerunvalidation$bd$lai~singlerunvalidation$bd$date, type = "l", main = "LAI", xlab = "Date", ylab = "LAI", col = 'DarkGreen')
plot(singlerunvalidation$bd$pet~singlerunvalidation$bd$date, type = "l", main = "Potential Evapotranspiration", xlab = "Date", ylab = "PET", col = 'DarkBlue')
plot(singlerunvalidation$bd$et~singlerunvalidation$bd$date, type = "l", main = "Evapotranspiration", xlab = "Date", ylab = "ET", col = 'darkslategray')

plot((singlerunvalidation$bd$unsat_stor/singlerunvalidation$bd$sat_def_z)~singlerunvalidation$bd$date, type = "l", main = "unsat_stor/sat_def_z", xlab = "Date", ylab = "vwc", col = 'DarkGreen')


```

Merge Observed and Simulated Data

```{r Merge Observed and Simulated Validation Datasets}
## merge values instead of subsetting to match up with simulated dates
singlerunvalidation$bd<- merge(singlerunvalidation$bd,obsws32, by.x = "date", by.y = "Date", all = FALSE)

## merge soil moisture values but include all data
singlerunvalidation$data <-merge(singlerunvalidation$bd,obsws32sm, by.x = "date", by.y = "Date", all = TRUE)
```

Plot Observed and Simulated Data, Plot RHESSys Outputs

```{r Plot Observed and Simulated Validation Datasets}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

## Plot Hydrograph comparing modeled and observed streamflow
{hydrograph(input=singlerunvalidation$bd, streamflow=singlerunvalidation$bd$observedstreamflow, streamflow2 = singlerunvalidation$bd$streamflow, timeSeries = singlerunvalidation$bd$date, precip = singlerunvalidation$bd$streamflow,
           P.units = "mm", S.units = "mm normalized by basin area", S1.col = 'Blue', S2.col = 'Red')

legend("topleft",inset = .2, legend=c("Observed Streamflow","Predicted Streamflow"), col=c("blue","red"), lty=1:2)}

{plot(singlerunvalidation$data$date,singlerunvalidation$data$observedstreamflow.x, type = "l", xlab = "Date", ylab = "Streamflow", main = "Predicted vs Observed Streamflow")
lines(singlerunvalidation$data$date, singlerunvalidation$data$streamflow, col = 'RED')
legend("topright",inset = 0, legend=c("Observed Streamflow","Predicted Streamflow"), col=c("black","red"), lty=1:1)}

{plot(singlerunvalidation$data$date,singlerunvalidation$data$observedstreamflow.x, type = "l", xlab = "Date", ylab = "Streamflow", main = "Predicted vs Observed Log Streamflow", log = "y")
lines(singlerunvalidation$data$date, singlerunvalidation$data$streamflow, col = 'RED')
legend("topright",inset = 0, legend=c("Observed Streamflow","Predicted Streamflow"), col=c("black","red"), lty=1:1)}

predictedvsobservedstreamlm<- lm(observedstreamflow~streamflow, data = singlerunvalidation$bd)

{plot(singlerunvalidation$bd$observedstreamflow,singlerunvalidation$bd$streamflow, xlab = "Observed", ylab = "Predicted", main = "Predicted vs Observed Streamflow", xlim = c(0,55), ylim = c(0,55))
abline(a=0, b=1, col = 'RED', lty = 2, lwd = 2)
legend("right",inset = 0.01, legend = paste("R2 =",format(summary(predictedvsobservedstreamlm)$r.squared,digits=3)))}

{plot(singlerunvalidation$data$date,(singlerunvalidation$data$rz_storage/singlerunvalidation$data$rootdepth), type = "l", ylim = c(0,0.35), xlab = "Date", ylab = "RZ_Storage/Rootdepth", main = "Predicted vs Observed Basin Scale Soil Moisture")
lines(singlerunvalidation$data$date, singlerunvalidation$data$mergedsoilmoisture, col = 'RED')
legend("topright",inset = 0, legend=c("Observed Soil Moisture","Predicted Soil Moisture"), col=c("black","red"), lty=1:1)}

predictedvsobservedsoillm<- lm((rz_storage/rootdepth)~mergedsoilmoisture, data = singlerunvalidation$data)

{plot((singlerunvalidation$data$rz_storage/singlerunvalidation$data$rootdepth),singlerunvalidation$data$mergedsoilmoisture, xlab = "Observed", ylab = "Predicted", main = "Predicted vs Observed Basin Scale Soil Moisture", xlim = c(0.05,0.4), ylim = c(0.05,0.4))
abline(a=0, b=1, col = 'RED', lty = 2, lwd = 2)
legend("right",inset = 0.01, legend = paste("R2 =",format(summary(predictedvsobservedsoillm)$r.squared,digits=3)))}

## Plot other model outputs

par(mfrow=c(5,1), mar = c(2,5,4,2))

plot(singlerunvalidation$bd$date,singlerunvalidation$bd$tmin, type = "l", col = 'BLUE', ylim = c(-15,30), ylab = "Temperature", xlab = "Date")
lines(singlerunvalidation$bd$date,singlerunvalidation$bd$tmax, type = "l", col = 'RED')
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$precip, type = "h", ylab = "Precipitation", xlab = "Date")
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$streamflow, type = "l", ylab = "Streamflow", xlab = "Date", col = 'blue')
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$baseflow, type = "l", ylab = "Baseflow", xlab = "Date", col = 'brown')
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$rz_storage, type = "l", ylab = "RZ Storage", xlab = "Date", col = 'dark green')
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$psn ,type = "l", ylab = "PSN", xlab = "Date")
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$et, type = "l", ylab = "Evaoptranspiration", xlab = "Date")
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$pet, type = "l", ylab = "PET", xlab = "Date")
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$gw.Qout, type = "l", ylab = "Q out", xlab = "Date")
plot(singlerunvalidation$bd$date,singlerunvalidation$bd$gw.storage, type = "l", ylab = "GW Storage", xlab = "Date")

par(mfrow=c(1,1))

### will only plot in growth mode , used to check that vegetation is initialized properly ##
plot(singlerunvalidation$bdg$date, singlerunvalidation$bdg$lai, type = "l", main = "LAI", xlab = "Date", col = 'dark green')
plot(singlerunvalidation$bdg$date, singlerunvalidation$bdg$soilc, type = "l", main = "Soil Carbon", xlab = "Date", col = 'brown')


## plot cal/val sets


obsws32<- read.csv(paste0(filedir,"input_files/streamflow/Daily Streamflow Watershed 32/Daily Streamflow Watershed 32/ObservedWS32.csv"))


obsws32$Date<- as.Date(obsws32$Date)

Caldates<- read.csv("outputfilter/caldates.csv")
Valdates<- read.csv("outputfilter/valdates.csv")

Caldates<- Caldates[,2]
Valdates<- Valdates[,2]

mergedsm<- read.table("outputfilter/mergedsm.csv", header = T, sep = ",")
obsws32sm<- read.table("outputfilter/obsws32sm.csv", header = T, sep = ",")
obsws32sm$Date<- as.Date(obsws32sm$Date,"%Y-%m-%d")

obsws32cal <- obsws32[obsws32$Date >= Caldates[1] & obsws32$Date <= Caldates[2], ]
obsws32val <- obsws32[obsws32$Date >= Valdates[1] & obsws32$Date <= Valdates[2], ]
obsws32smcal <- obsws32sm[obsws32sm$Date >= Caldates[1] & obsws32sm$Date <= Caldates[2], ]
obsws32smval <- obsws32sm[obsws32sm$Date >= Valdates[1] & obsws32sm$Date <= Valdates[2], ]


{plot(singlerunvalidation$data$date,(singlerunvalidation$data$rz_storage/singlerunvalidation$data$rootdepth), type = "l", ylim = c(0,0.35), xlab = "Date", ylab = "RZ_Storage/Rootdepth", main = "Predicted vs Observed Basin Scale Soil Moisture")
lines(obsws32smcal$Date,obsws32smcal$mergedsoilmoisture, col = "BLUE")
lines(obsws32smval$Date,obsws32smval$mergedsoilmoisture, col = "RED")
legend("topright",inset = 0, legend=c("Predicted Soil Moisture","Observed Calibration Soil Moisture", "Observed Validation Soil Moisture"), col=c("black","blue","red"), lty=c(1,1,1))}

{plot(singlerunvalidation$data$date,singlerunvalidation$data$streamflow, type = "l", xlab = "Date", ylab = "Streamflow", main = "Predicted vs Observed Basin Scale Streamflow")
lines(obsws32cal$Date,obsws32cal$observedstreamflow, col = "BLUE")
lines(obsws32val$Date,obsws32val$observedstreamflow, col = "RED")
legend("topright",inset = 0, legend=c("Predicted Streamflow","Observed Calibration Streamflow", "Observed Validation Streamflow"), col=c("black","blue","red"), lty=c(1,1,1))}
```

Spatial Data import and display of inputs

```{r Display Spatial Inputs and Simulated Validation Soil Moisture, fig.height=10, fig.width=14}
# Original input maps have been cropped before processing in R, without cropping the RHESSYs preparation in the R environment will not function properly, below are the input maps used by R
setwd(system.file("extdata/", package = "RHESSysIOinR"))

ws32<- raster("grassexport/cwt_ws32/basin_ws32.tif")
patches<- raster("grassexport/cwt_ws32/patch.tif")
acc10<- raster("grassexport/cwt_ws32/acc10.tif")
aspect10<- raster("grassexport/cwt_ws32/aspect10.tif")
#bt1000<- raster("C:/Users/Carlos/Documents/R/win-library/4.0/RHESSysIOinR/extdata/rasters/cwt_ws32/ws32/b.t1000.tif")
dem10<- raster("grassexport/cwt_ws32/dem10.tif")
dem10f<- raster("grassexport/cwt_ws32/dem10f.tif")
dir10<- raster("grassexport/cwt_ws32/dir10.tif")
drain10<- raster("grassexport/cwt_ws32/drain10.tif")
#ht1000<- raster("C:/Users/Carlos/Documents/R/win-library/4.0/RHESSysIOinR/extdata/rasters/cwt_ws32/ws32/h.t1000.tif")
hillslope<- raster("grassexport/cwt_ws32/hillslope.tif")
roads<- brick("grassexport/cwt_ws32/roads.tif")
#shade10<- raster("C:/Users/Carlos/Documents/R/win-library/4.0/RHESSysIOinR/extdata/rasters/cwt_ws32/ws32/shade10.tif")
slope10<- raster("grassexport/cwt_ws32/slope10.tif")
streams<- raster("grassexport/cwt_ws32/streams.tif")
topidx10<- raster("grassexport/cwt_ws32/topidx10_100.tif")
#xmap<- raster("C:/Users/Carlos/Documents/R/win-library/4.0/RHESSysIOinR/extdata/rasters/cwt_ws32/ws32/xmap.tif")
#ymap<- raster("C:/Users/Carlos/Documents/R/win-library/4.0/RHESSysIOinR/extdata/rasters/cwt_ws32/ws32/ymap.tif")

ObservedSMLocationsnt <- read_sf(paste0(filedir,"input_files/soil_moisture/knb-lter-cwt.1308.19/1308.kml"))

ObservedSMLocations <- st_transform(ObservedSMLocationsnt, crs(dem10))

plot(ws32, main ="ws32")
plot(patches, main = "patches")
plot(acc10, main = "acc10")
plot(aspect10, main = "aspect10")
#levelplot(bt1000, col.regions = brewer.pal(name='Spectral', n = 11),at=seq(0,6500,1))
#plot(bt1000, main = "bt1000")
plot(dem10, main = "dem10")
plot(dem10f, main = "dem10f")
plot(dir10, main = "dir10")
plot(drain10, main = "drain10")
#plot(ht1000, main = "ht1000")
{plot(hillslope, main = "hillslope")
  points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", lwd = 1)}
plot(roads, main = "roads")
#plot(shade10, main = "shade10")
plot(slope10, main = "slope10")
plot(streams, main = "streams")
plot(topidx10, main = "topidx10")
#plot(xmap, main = "xmap")
#plot(ymap, main = "ymap")

```

Show locations of measured soil moisture and soil maps

```{r Display Locations of Observed Soil Moisture, fig.height=10, fig.width=10}
setwd(system.file("extdata/", package = "RHESSysIOinR"))


soilstaticmap<- raster("grassexport/cwt_ws32/staticsoil.tif")

soilssurgo<- raster("grassexport/cwt_ws32/ssurgosoil.tif")
soildsm<- raster("grassexport/cwt_ws32/dsmsoil.tif")

colorramp<- viridis(128)

{plot(dem10, col = colorramp, main = "Coweeta Observed Soil Moisture Locations")
points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", lwd = 1)}


matrix(c(ObservedSMLocations$Name[3],ObservedSMLocations$Name[4],ObservedSMLocations$Name[5],ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1],ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), nrow = 3, ncol = 3)


{plot(soilstaticmap, col = colorramp, main = "Coweeta Observed Soil Moisture Locations on Static Soil Map")
points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", lwd = 1)}

{plot(soilssurgo, col = colorramp, main = "Coweeta Observed Soil Moisture Locations on SSURGO Soil Map")
points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", lwd = 1)}

{plot(soildsm, col = colorramp, main = "Coweeta Observed Soil Moisture Locations on RSS Soil Map")
points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", lwd = 1)}


```

Display outputs from model run

```{r Display Spatial Outputs and Simulated Validation Soil Moisture, fig.height=10, fig.width=14}
setwd(system.file("extdata/", package = "RHESSysIOinR"))
## Extract values for single date
displayday <- singlerunvalidation$pd[which(singlerunvalidation$pd$date=="2018-03-21")]

## reclassify all patches to na and then reclassify to model output values
reclass_all<- c(-2147483648, 2147483647,NA)
reclass_allm<- matrix(reclass_all, ncol = 3, byrow = TRUE)
displaydaypatches<- reclassify(patches,reclass_allm)

displayday$calculatedrzsm <- displayday$rz_storage/displayday$root.depth

reclass_displayday <- cbind(displayday$patchID,displayday$calculatedrzsm)
reclass_displayday <- as.matrix(reclass_displayday)

displaydaypatches<- reclassify(patches,reclass_displayday)

## Plot Spatial Data Output for RZ Storage
#plot(displaydaypatches, xlim = c(280900,282500), ylim = c(4870000,4872000))

{plot(mask(displaydaypatches,ws32), xlim = c(275000,276200), ylim = c(3881000,3881600))
title("CW WS32 RZ Soil Moisture Prediction for 1 day", line = -2)
points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", cex = 2)
scalebar(200, xy=c(275200, 3880800), type='line',divs = "2")}


reclass_displayday <- cbind(displayday$patchID,displayday$root.depth)
reclass_displayday <- as.matrix(reclass_displayday)
displaydaypatches2<- reclassify(patches,reclass_displayday)

{plot(mask(displaydaypatches2,ws32), xlim = c(275000,276200), ylim = c(3881000,3881600))
title("CW WS32 root.depth Prediction for 1 day", line = -2)
scalebar(200, xy=c(275200, 3880800), type='line',divs = "2")}


reclass_displayday <- cbind(displayday$patchID,displayday$rz_storage)
reclass_displayday <- as.matrix(reclass_displayday)
displaydaypatches3<- reclassify(patches,reclass_displayday)

{plot(mask(displaydaypatches3,ws32), xlim = c(275000,276200), ylim = c(3881000,3881600))
title("CW WS32 rz_storage Prediction for 1 day", line = -2)
scalebar(200, xy=c(275200, 3880800), type='line',divs = "2")}

reclass_displayday <- cbind(displayday$patchID,displayday$sat_def)
reclass_displayday <- as.matrix(reclass_displayday)
displaydaypatches4<- reclassify(patches,reclass_displayday)

{plot(mask(displaydaypatches4,ws32), xlim = c(275000,276200), ylim = c(3881000,3881600))
title("CW WS32 Saturation Deficit Prediction for 1 day in mm", line = -2)
scalebar(200, xy=c(275200, 3880800), type='line',divs = "2")}

reclass_displayday <- cbind(displayday$patchID,(displayday$unsat_stor/displayday$sat_def_z))
reclass_displayday <- as.matrix(reclass_displayday)
displaydaypatches5<- reclassify(patches,reclass_displayday)

{plot(mask(displaydaypatches5,ws32), xlim = c(275000,276200), ylim = c(3881000,3881600))
title("CW WS32 unsaturated soil moisture vwc Prediction for 1 day", line = -2)
scalebar(200, xy=c(275200, 3880800), type='line',divs = "2")}


reclass_displayday <- cbind(displayday$patchID,((displayday$root_zone.S*displayday$potential_rz_store)/displayday$root.depth))
reclass_displayday <- as.matrix(reclass_displayday)
displaydaypatches6<- reclassify(patches,reclass_displayday)

{plot(mask(displaydaypatches6,ws32), xlim = c(275000,276200), ylim = c(3881000,3881600), zlim = c(0,0.50))
title("CW WS32 Rooting Depth VWC Prediction for 1 day", line = -2)
points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", cex = 2)
scalebar(200, xy=c(275200, 3880800), type='line',divs = "2")}

{plot(mask(displaydaypatches,ws32), xlim = c(275000,276200), ylim = c(3881000,3881600), zlim = c(0,0.50))
title("CW WS32 RZ Soil Moisture Prediction for 1 day", line = -2)
points(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1]), c(ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), pch = 21, col = "black", bg = "red", cex = 2)
scalebar(200, xy=c(275200, 3880800), type='line',divs = "2")}

## Plot Hydrograph
{hydrograph(input=singlerunvalidation$bd, streamflow=singlerunvalidation$bd$observedstreamflow, streamflow2 = singlerunvalidation$bd$streamflow, timeSeries = singlerunvalidation$bd$date, precip = singlerunvalidation$bd$streamflow,
           P.units = "mm", S.units = "mm normalized by basin area", S1.col = 'Blue', S2.col = 'Red')
legend("topleft",inset = .2, legend=c("Observed Streamflow","Predicted Streamflow"), col=c("blue","red"), lty=1:2)}

### end show patch map

patchnumbers<- raster::extract(patches,matrix(c(ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1],ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2]), nrow = 3, ncol = 2))


ObserveSoilMoisturePatches <- matrix(c(ObservedSMLocations$Name[3],ObservedSMLocations$Name[4],ObservedSMLocations$Name[5],ObservedSMLocations[[3]][[3]][1],ObservedSMLocations[[3]][[4]][1],ObservedSMLocations[[3]][[5]][1],ObservedSMLocations[[3]][[3]][2],ObservedSMLocations[[3]][[4]][2],ObservedSMLocations[[3]][[5]][2],patchnumbers), nrow = 3, ncol = 4)


## Extract values for a single patch, select patch ID
ObserveSoilMoisturePatches[1,4]
displaypatch<- singlerunvalidation$pd[which(singlerunvalidation$pd$patchID==ObserveSoilMoisturePatches[1,4])]
ObserveSoilMoisturePatches[2,4]
displaypatch2<- singlerunvalidation$pd[which(singlerunvalidation$pd$patchID==ObserveSoilMoisturePatches[2,4])]
ObserveSoilMoisturePatches[3,4]
displaypatch3<- singlerunvalidation$pd[which(singlerunvalidation$pd$patchID==ObserveSoilMoisturePatches[3,4])]

plot(displaypatch$date,(displaypatch$rz_storage/displaypatch$root.depth), type = "l", xlab = "Date", ylab = "RZ Storage", main = 'CW WS27 More Detailed RZ Storage Patch #147322', ylim = c(0,.4))

head(obsws32cal)
head(obsws32val)
head(obsws32smcal)
head(obsws32smval)

#subset Validation range based on Val Dates set by available SM data

validationsubset <- singlerunvalidation$bd[singlerunvalidation$bd$date >= Valdates[1] & singlerunvalidation$bd$date <= Valdates[2], ]

## Fit tests for streamflow

NSE(validationsubset$streamflow,validationsubset$observedstreamflow)

NSE(validationsubset$streamflow,validationsubset$observedstreamflow, FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE)

KGE(validationsubset$streamflow,validationsubset$observedstreamflow)

# Plot Observed and Predicted Streamflow

{plot(validationsubset$date,validationsubset$streamflow, type = 'l', lty = 2, col = 'RED', main = "Validation Subset Observed and Predicted Streamflow")
lines(validationsubset$date,validationsubset$observedstreamflow, type = 'l', col = 'BLACK')
legend("topright",inset = .1, legend=c("Observed Streamflow","Predicted Streamflow"), col=c("black","red"), lty=1:2)}


{plot(singlerunvalidation$bd$date,singlerunvalidation$bd$streamflow, type = 'l', lty = 2, col = 'RED', main = "Observed and Predicted Streamflow")
lines(singlerunvalidation$bd$date,singlerunvalidation$bd$observedstreamflow, type = 'l', col = 'BLACK')
legend("topright",inset = .1, legend=c("Observed Streamflow","Predicted Streamflow"), col=c("black","red"), lty=1:2)}

#Plot Model RZ SM outputs for patch
{plot(displaypatch$date,(displaypatch$rz_storage/displaypatch$root.depth), type = "l",ylim = c(0,0.50), ylab = "Root Zone Storage (mm/m) ", xlab = "Date", main = 'Root Zone Display Patch')
lines(displaypatch$date,(displaypatch$rz_field_capacity/displaypatch$root.depth), type = "l", col = "blue")
lines(displaypatch$date,(displaypatch$rz_wilting_point/displaypatch$root.depth), type = "l", col = "red")
legend("bottomleft",inset = .1, legend=c("Predicted RZ Storage","RZ Field Capacity","RZ Wilting Point"), col=c("black","blue","red"), lty=1:1)}


{plot(displaypatch$date,(displaypatch$unsat_stor/displaypatch$sat_def_z), type = "l",ylim = c(0,0.50), ylab = "Root Zone Storage (mm/m) ", xlab = "Date", main = 'Root Zone Display Patch')
legend("bottomleft",inset = .1, legend=c("Predicted RZ Storage","RZ Field Capacity","RZ Wilting Point"), col=c("black","blue","red"), lty=1:1)}

# must take seperate markdown file into account, load in mean sm files

# {plot(smws32_2mean$Group.1,smws32_2mean$smois30A, type = "l", ylim = c(0,0.4), col = 'BLUE', xlim=as.Date(c("2016-10-1","2019-10-1")), main = "CW WS32 More Saturated Zone VWC vs Patch #154914",
#      xlab = "Date", ylab = "Soil Moisture mm")
# lines(displaypatch$date,(displaypatch$unsat_stor/displaypatch$sat_def_z), type = "l", xlab = "Date", ylab = "RZ Storage", main = 'CW WS27 More Detailed RZ Storage Patch #154914', col = 'red')
# legend("bottomleft",inset = .01, legend=c("Observed Soil Moisture","Predicted Soil Moisture"), col=c("blue","red"), lty=1:1)}

```

Combine calibration outputs

```{r}

setwd(system.file("extdata/", package = "RHESSysIOinR"))
getwd()

dsmdist<- read.csv("out/dsmcalibrationoutput.txt", sep=" ")
ssurgodist<- read.csv("out/ssurgocalibrationoutput.txt", sep = " ")
staticdist<- read.csv("out/staticcalibrationoutput.txt", sep = " ")

paged_table(ssurgodist)
paged_table(dsmdist)
paged_table(staticdist)

ssurgodist$treatment<- 'SSURGO'
dsmdist$treatment<- 'RSS'
staticdist$treatment<- 'Static'

combineddist<- rbind(ssurgodist, dsmdist, staticdist)

ggplot(combineddist, aes(calsmKGE, fill = treatment))+ geom_density(alpha=0.3)

ggplot(combineddist, aes(calKGE, fill = treatment))+ geom_density(alpha=0.3)


smKGEplot<- ggplot(combineddist, aes(calsmKGE, y = treatment, fill = treatment))+ geom_boxplot()+ xlab("Soil Moisture KGE - Calibration") +ylab("Soil Map Input")

KGEplot<- ggplot(combineddist, aes(calKGE, y = treatment, fill = treatment))+ geom_boxplot()+ xlab("Streamflow KGE - Calibration")+ylab(NULL)

ggarrange(smKGEplot,KGEplot, common.legend = TRUE, legend = "top")


smKGEplot<- ggplot(combineddist, aes(calsmKGE, y = treatment, fill = treatment))+ geom_boxplot()+ xlab("Soil Moisture KGE - Calibration") +ylab("Soil Map Input")

KGEplot<- ggplot(combineddist, aes(calKGE, y = treatment, fill = treatment))+ geom_boxplot()+ xlab("Streamflow KGE - Calibration")+ylab(NULL)

ggarrange(smKGEplot,KGEplot, common.legend = TRUE, legend = "top")


text <- "Coweeta Watershed 32 Calibration Performance"
# Create a text grob
tgrob <- text_grob(text,size = 16)
# Draw the text
plot_0 <- as_ggplot(tgrob) + theme(plot.margin = margin(0,3,0,12, "cm"))

ggarrange(plot_0,NULL,smKGEplot,KGEplot,
          ncol = 2,nrow = 2,heights = c(1,5), common.legend = TRUE, legend = "none")

smKGEviolinplot<- ggplot(combineddist, aes(calsmKGE, y = treatment, fill = treatment))+ geom_violin()+ xlab("Soil Moisture KGE - Calibration") +ylab("Soil Map Input")

KGEviolinplot<- ggplot(combineddist, aes(calKGE, y = treatment, fill = treatment))+ geom_violin()+ xlab("Streamflow KGE - Calibration")+ylab(NULL)

ggarrange(smKGEviolinplot,KGEviolinplot, common.legend = TRUE, legend = "top")

library(multcompView)
library(datasets)

oneway.test(valsmKGE~treatment, data= combineddist, var.equal = TRUE)

anova<- aov(valsmKGE~treatment, data = combineddist)

summary(anova)

tukey <- TukeyHSD(anova)
print(tukey)

cld <- multcompLetters4(anova,tukey)

print(cld)

```

Run top n model runs for validation

50/50 weighting

validation will only run 2 years before prediction time period

```{r}
combineddist

plot(ssurgodist$calKGE, ssurgodist$calsmKGE, pch = 19, xlab = "Streamflow KGE", ylab = "Soil Moisture KGE")

combineddist$weightedcal<- rowMeans(combineddist[,c("calKGE","calsmKGE")],)

ordereddist <- combineddist[order(-combineddist$weightedcal),]

ordereddist[ordereddist$treatment=="Static",]

topstatic<- ordereddist[ordereddist$treatment=="Static",]
topssurgo<- ordereddist[ordereddist$treatment=="SSURGO",]
toprss<- ordereddist[ordereddist$treatment=="RSS",]

toprss

```
RSS - prepare to run model n times for patch 1 - only for validation time series

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))
soildefs <- list.files("defs/", pattern = "soil_", full.names = TRUE)

getwd()


validationruns <- 100

input_rhessys = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32dsm.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32dsm",
  flowtable = "CWWS32dsm.flow",
  start = "2015 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32valrss1",
commandline_options = c(""))

input_tec_data = IOin_tec_std(start = "2017 11 1 1",
                              end = "2018 11 1 1",
                              output_state = FALSE)

input_hdr = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = "defs/lu_undev.def",
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")

```
RSS - Run RHESSYs

```{r Run RHESSYS in Parallel using foreach RSS 1}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

n.cores <- parallel::detectCores() - 4  
  
my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
doParallel::registerDoParallel(cl = my.cluster)

start_time = Sys.time()
  
  foreach(i = 1:validationruns, .packages = 'RHESSysIOinR') %dopar% {
    
    stdpars<- IOin_cmd_pars(m = toprss[i,]$m,
                            k = toprss[i,]$k,
                            soil_dep= toprss[i,]$soil_dep,
                            m_v = toprss[i,]$m_v,
                            k_v = toprss[i,]$k_v,
                            gw1 = toprss[i,]$gw1,
                            gw2 = toprss[i,]$gw2,
                            pa = toprss[i,]$pa,
                            po = toprss[i,]$po,
                            vgseng1 = toprss[i,]$vgseng1,
                            vgseng2 = toprss[i,]$vgseng2,
                            vgseng3 = toprss[i,]$vgseng3)
    
    run_rhessys_single(input_rhessys = input_rhessys,
                       hdr_files = input_hdr,
                       tec_data = input_tec_data,
                       cmd_pars = stdpars,
                       output_filter = c(patchfilterpatch1,patchfilterpatch2,patchfilterpatch3),
                       runID = i)
  } 
  
  end_time = Sys.time()
  end_time - start_time
  
```
##prepare data
RSS - prepare site specific soil moisture datasets, make sure that each patch looks at this
Site 2 at 60cm A was removed from the averaging because of no data

```{r}
#Prepare observed Soil moisture data
obsws32smvalclean<- data.frame(obsws32smval$Date,obsws32smval$mergedsoilmoisture)
names(obsws32smvalclean)<- c('Date','mergedsoilmoisture')

obsws32smvalclean

obsws32smvalclean$Site3<- rowMeans(subset(obsws32smval, select = c(smois30site3a,smois60site3a,smois30site3b,smois60site3b), na.rm = TRUE))
obsws32smvalclean$Site2<- rowMeans(subset(obsws32smval, select = c(smois30site2a,smois30site2b,smois60site2b), na.rm = TRUE))
obsws32smvalclean$Site1<- rowMeans(subset(obsws32smval, select = c(smois30site1a,smois60site1a,smois30site1b,smois60site1b), na.rm = TRUE))

obsws32smvalclean

plot(obsws32smvalclean$Date,obsws32smvalclean$Site3, type = "l", main = "SITE 3 Averaged Sensor Output")

obsws32smvalclean$Site3


plot(obsws32smvalclean$Date, 
     obsws32smvalclean$mergedsoilmoisture, 
     type = "l", 
     main = "All Site Averaged Sensor Output", 
     lty = 3, 
     lwd = 2, 
     ylim = c(0.10, 0.35), 
     xlab = "Date", 
     ylab = "Soil Moisture")

lines(obsws32smvalclean$Date, obsws32smvalclean$Site1, col = 'blue', lwd = 1)
lines(obsws32smvalclean$Date, obsws32smvalclean$Site2, col = 'red', lwd = 1)
lines(obsws32smvalclean$Date, obsws32smvalclean$Site3, col = 'green', lwd = 1)

legend("topright", 
       legend = c("Averaged Output", "Site 1", "Site 2", "Site 3"), 
       col = c("black", "blue", "red", "green"), 
       lty = c(3, 1, 1, 1), 
       lwd = 2)


```
# must modify vwc calculation to use * instead of /
RSS - read 20 runs, create table and append to table for site 1

```{r}
# Read in RHESSys Validation runs

setwd(system.file("extdata/", package = "RHESSysIOinR"))


#create rootzonevwc column
for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch1_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valrss1_run",i),temp_dataset)
}

##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valrss1_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}


valmergesm1


validationsmNSElistsite1<- c()
validationsmlnNSElistsite1<- c()
validationsmKGElistsite1<- c()


#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site1"))))))
  validationsmKGElistsite1[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }


validationsmKGElistsite1


```
RSS - Explore Site 1

```{r}

'Exploring Site 1 RSS'


cwws32valrss1_run1

{plot(cwws32valrss1_run1$date, cwws32valrss1_run1$rootzonevwc, type = 'l', ylim = c(0.12,0.26))
lines(valmergesm1$Date,valmergesm1$Site1, col = 'BLUE', lty = 3)}



{plot(cwws32valrss1_run3$date, cwws32valrss1_run3$rootzonevwc, type = 'l', ylim = c(0.12,0.26))
lines(valmergesm3$Date,valmergesm3$Site1, col = 'BLUE', lty = 3)}


plot((valmergesm3$rootzone.S*valmergesm3$rootzone.potential_sat), type = "l")
plot(valmergesm3$Date, (valmergesm3$rz_storage), type = "l")
plot(valmergesm3$Date, valmergesm3$Site1, type = "l")


{plot(cwws32valrss1_run3$date, cwws32valrss1_run3$rootzonevwc, type = 'l', ylim = c(0.0,0.26))
lines(valmergesm3$Date,valmergesm3$Site1, col = 'BLUE', lty = 3)}


{plot(cwws32valrss1_run2$date, cwws32valrss1_run2$rootzonevwc, type = 'l', ylim = c(0.0,0.26))
lines(valmergesm2$Date,valmergesm2$Site1, col = 'BLUE', lty = 3)}

paste0("valmergesm",i,"$rootzonevwc")


{plot(cwws32valrss1_run3$date, (cwws32valrss1_run3$rootzone.S*cwws32valrss1_run3$rootzone.potential_sat), type = 'l', ylim = c(0.0,0.26))
lines(valmergesm3$Date,valmergesm3$Site1, col = 'BLUE', lty = 3)}

# {plot(cwws32valrss1_run44$date, (cwws32valrss1_run44$rootzone.S/cwws32valrss1_run44$root.depth), type = 'l', ylim = c(0.0,0.26))
# lines(valmergesm4$Date,valmergesm4$Site1, col = 'BLUE', lty = 3)}
# 
# {plot(cwws32valrss1_run43$date, (cwws32valrss1_run43$rootzone.S/cwws32valrss1_run43$rootzone.potential_sat), type = 'l', ylim = c(0.0,0.26))
# lines(valmergesm1$Date,valmergesm1$Site1, col = 'BLUE', lty = 3)}

```
RSS - prepare to run model n times for patch 2 - only for validation time series

RSS - run model in parallel 

RSS - read n runs, create table and append to table

```{r}

setwd(system.file("extdata/", package = "RHESSysIOinR"))
# Read in RHESSys Validation runs
for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch2_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valrss2_run",i),temp_dataset)
}

##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valrss2_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}

valmergesm1

validationsmNSElistsite2<- c()
validationsmlnNSElistsite2<- c()
validationsmKGElistsite2<- c()

#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site2"))))))
  validationsmKGElistsite2[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }

validationsmKGElistsite2

```
RSS - Explore Site 2
## not updated 6/27
```{r}
# 
# 'Exploring Site 2 RSS'
# 
# 
# cwws32valrss2_run1
# 
# {plot(cwws32valrss2_run1$bd$date, (cwws32valrss2_run1$bd$rz_storage/cwws32valrss2_run1$bd$rootdepth), type = 'l', ylim = c(0.12,0.26))
# lines(valmergesm1$Date,valmergesm1$Site2, col = 'BLUE', lty = 3)}
# 
# 
# 
# {plot(cwws32valrss2_run3$bd$date, (cwws32valrss2_run3$bd$rz_storage/cwws32valrss2_run3$bd$rootdepth), type = 'l', ylim = c(0.12,0.26))
# lines(valmergesm3$Date,valmergesm3$Site2, col = 'BLUE', lty = 3)}
# 
# paste0("valmergesm",i,"$rootzonevwc")
# 
# plot((valmergesm3$root_zone.S*valmergesm3$potential_rz_store)/(valmergesm3$root.depth), type = "l")
# plot(valmergesm3$Date, (valmergesm3$rz_storage/valmergesm3$root.depth), type = "l")
# plot(valmergesm3$Date, valmergesm3$Site2, type = "l")
# 
# 
# {plot(cwws32valrss2_run3$pd$date, (cwws32valrss2_run3$pd$rz_storage/cwws32valrss2_run3$pd$root.depth), type = 'l', ylim = c(0.0,0.26))
# lines(valmergesm3$Date,valmergesm3$Site2, col = 'BLUE', lty = 3)}
# 
# 
# {plot(cwws32valrss2_run3$pd$date, ((cwws32valrss2_run3$pd$root_zone.S*cwws32valrss2_run3$pd$potential_rz_store)/cwws32valrss2_run3$pd$root.depth), type = 'l', ylim = c(0.0,0.26))
# lines(valmergesm3$Date,valmergesm3$Site2, col = 'BLUE', lty = 3)}
# 
# paste0("valmergesm",i,"$rootzonevwc")
# 
# 
# {plot(cwws32valrss2_run3$bd$date, (cwws32valrss2_run3$bd$rz_storage/cwws32valrss2_run3$bd$rootdepth), type = 'l', ylim = c(0.12,0.26))
# lines(valmergesm3$Date,valmergesm3$Site2, col = 'BLUE', lty = 3)}
# 
# {plot(cwws32valrss2_run43$pd$date, (cwws32valrss2_run43$pd$rz_storage/cwws32valrss2_run43$pd$root.depth), type = 'l', ylim = c(0.0,0.26))
# lines(valmergesm43$Date,valmergesm43$Site2, col = 'BLUE', lty = 3)}
# 
# {plot(cwws32valrss2_run43$bd$date, (cwws32valrss2_run43$bd$rz_storage/cwws32valrss2_run43$bd$rootdepth), type = 'l', ylim = c(0.12,0.26))
# lines(valmergesm1$Date,valmergesm1$Site2, col = 'BLUE', lty = 3)}

```
prepare to run model 20 times for patch 3 - only for validation time series

RSS - run model in parallel

RSS - read n runs, create table and append to table

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

# Read in RHESSys Validation runs
for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch3_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valrss3_run",i),temp_dataset)
}
##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valrss3_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}


validationsmNSElistsite3<- c()
validationsmlnNSElistsite3<- c()
validationsmKGElistsite3<- c()

#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site3"))))))
  validationsmKGElistsite3[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }


validationsmKGElistsite3

valmergesm2

 
```
RSS - Explore Site 3
#not updated 6/27
```{r}
# 
# 'Exploring Site 3 RSS'
# 
# 
# cwws32valrss3_run1
# 
# {plot(cwws32valrss3_run1$bd$date, (cwws32valrss3_run1$bd$rz_storage/cwws32valrss3_run1$bd$rootdepth), type = 'l', ylim = c(0.12,0.26))
# lines(valmergesm1$Date,valmergesm1$Site3, col = 'BLUE', lty = 3)}
# 
# 
# 
# {plot(cwws32valrss3_run3$bd$date, (cwws32valrss3_run3$bd$rz_storage/cwws32valrss3_run3$bd$rootdepth), type = 'l', ylim = c(0.12,0.26))
# lines(valmergesm3$Date,valmergesm3$Site3, col = 'BLUE', lty = 3)}
# 
# paste0("valmergesm",i,"$rootzonevwc")
# 
# plot((valmergesm3$root_zone.S*valmergesm3$potential_rz_store)/(valmergesm3$root.depth), type = "l")
# plot(valmergesm3$Date, (valmergesm3$rz_storage/valmergesm3$root.depth), type = "l")
# plot(valmergesm3$Date, valmergesm3$Site3, type = "l")
# 
# 
# {plot(cwws32valrss3_run3$pd$date, (cwws32valrss3_run3$pd$rz_storage/cwws32valrss3_run3$pd$root.depth), type = 'l', ylim = c(0.0,0.45))
# lines(valmergesm3$Date,valmergesm3$Site3, col = 'BLUE', lty = 3)}
# 
# 
# {plot(cwws32valrss3_run3$pd$date, ((cwws32valrss3_run3$pd$root_zone.S*cwws32valrss3_run3$pd$potential_rz_store)/cwws32valrss3_run3$pd$root.depth), type = 'l', ylim = c(0.0,0.45))
# lines(valmergesm3$Date,valmergesm3$Site3, col = 'BLUE', lty = 3)}
# 
# ## based on landscape position the previous two figures differ, the second figure does not count inundation as "0" VWC
# 
# #paste0("valmergesm",i,"$rootzonevwc")
# 
# 
# {plot(cwws32valrss3_run3$bd$date, (cwws32valrss3_run3$bd$rz_storage/cwws32valrss3_run3$bd$rootdepth), type = 'l', ylim = c(0.0,0.45))
# lines(valmergesm3$Date,valmergesm3$Site3, col = 'BLUE', lty = 3)}
# 
# {plot(cwws32valrss3_run43$pd$date, (cwws32valrss3_run43$pd$rz_storage/cwws32valrss3_run43$pd$root.depth), type = 'l', ylim = c(0.0,0.35))
# lines(valmergesm43$Date,valmergesm43$Site3, col = 'BLUE', lty = 3)}
# 
# {plot(cwws32valrss3_run43$bd$date, (cwws32valrss3_run43$bd$rz_storage/cwws32valrss3_run43$bd$rootdepth), type = 'l', ylim = c(0.12,0.26))
# lines(valmergesm1$Date,valmergesm1$Site3, col = 'BLUE', lty = 3)}

```
RSS - merge all 3 sites

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

threesitevalidation1<- data.frame(cbind(site1=as.numeric(validationsmKGElistsite1),site2=as.numeric(validationsmKGElistsite2),site3=as.numeric(validationsmKGElistsite3)))


threesitevalidation<-threesitevalidation1

threesitevalidation$allsites<- rowMeans(threesitevalidation1,na.rm=FALSE)

threesitevalidation


boxplot(threesitevalidation1, main = "Boxplot of Patch Specific Soil Moisture KGE - RSS")
boxplot(threesitevalidation, main = "Boxplot of Patch Specific Soil Moisture KGE - RSS")



plot(validationsmKGElistsite1,validationsmKGElistsite3)

write.csv(threesitevalidation,'threesitevalidationrss.csv')


```
STATIC

run model n times for patch 1 - only for validation time series

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

soildefs <- list.files("defs/", pattern = "soil_", full.names = TRUE)

input_rhessys = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32static.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32static",
  flowtable = "CWWS32static.flow",
  start = "2015 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32valstatic1",
commandline_options = c(""))

input_tec_data = IOin_tec_std(start = "2017 11 1 1",
                              end = "2018 11 1 1",
                              output_state = FALSE)

input_hdr = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = "defs/lu_undev.def",
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")

```
run static

```{r Run RHESSYS in Parallel using foreach STATIC 1}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

n.cores <- parallel::detectCores() - 4  
  
my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
doParallel::registerDoParallel(cl = my.cluster)

start_time = Sys.time()
  
  foreach(i = 1:validationruns, .packages = 'RHESSysIOinR') %dopar% {
    
    stdpars<- IOin_cmd_pars(m = topstatic[i,]$m,
                            k = topstatic[i,]$k,
                            soil_dep= topstatic[i,]$soil_dep,
                            m_v = topstatic[i,]$m_v,
                            k_v = topstatic[i,]$k_v,
                            gw1 = topstatic[i,]$gw1,
                            gw2 = topstatic[i,]$gw2,
                            pa = topstatic[i,]$pa,
                            po = topstatic[i,]$po,
                            vgseng1 = topstatic[i,]$vgseng1,
                            vgseng2 = topstatic[i,]$vgseng2,
                            vgseng3 = topstatic[i,]$vgseng3)
    
    run_rhessys_single(input_rhessys = input_rhessys,
                       hdr_files = input_hdr,
                       tec_data = input_tec_data,
                       cmd_pars = stdpars,
                       output_filter = c(patchfilterpatch1,patchfilterpatch2,patchfilterpatch3),
                       runID = i)
  } 
  
  end_time = Sys.time()
  end_time - start_time
```
read 20 runs, create table and append to table for patch 1

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

# Read in RHESSys Validation runs

for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch1_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valstatic1_run",i),temp_dataset)
}

##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valstatic1_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}

valmergesm1

validationsmNSElistsite1<- c()
validationsmlnNSElistsite1<- c()
validationsmKGElistsite1<- c()


#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site1"))))))
  validationsmKGElistsite1[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }

validationsmKGElistsite1

```
prepare to run model 20 times for patch 2 - only for validation time series

run model in parallel 

read 20 runs, create table and append to table

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

# Read in RHESSys Validation runs

for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch2_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valstatic2_run",i),temp_dataset)
}

##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valstatic2_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}


valmergesm1


validationsmNSElistsite2<- c()
validationsmlnNSElistsite2<- c()
validationsmKGElistsite2<- c()


#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site2"))))))
  validationsmKGElistsite2[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }


validationsmKGElistsite2



```

prepare to run model 20 times for patch 3 - only for validation time series

run model in parallel

read 20 runs, create table and append to table

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

# Read in RHESSys Validation runs
for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch3_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valstatic3_run",i),temp_dataset)
}


##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valstatic3_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}


validationsmNSElistsite3<- c()
validationsmlnNSElistsite3<- c()
validationsmKGElistsite3<- c()

#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site3"))))))
  validationsmKGElistsite3[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }


validationsmKGElistsite3

valmergesm2
 
```
merge all 3 sites 

```{r}

setwd(system.file("extdata/", package = "RHESSysIOinR"))

threesitevalidation1<- data.frame(cbind(site1=as.numeric(validationsmKGElistsite1),site2=as.numeric(validationsmKGElistsite2),site3=as.numeric(validationsmKGElistsite3)))


threesitevalidation<-threesitevalidation1

threesitevalidation$allsites<- rowMeans(threesitevalidation1,na.rm=FALSE)

threesitevalidation


boxplot(threesitevalidation1, main = "Boxplot of Patch Specific Soil Moisture KGE - Static")
boxplot(threesitevalidation, main = "Boxplot of Patch Specific Soil Moisture KGE - Static")



plot(validationsmKGElistsite1,validationsmKGElistsite3)

write.csv(threesitevalidation,'threesitevalidationstatic.csv')


```
SSURGO

run model 20 times for patch 1 - only for validation time series

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

soildefs <- list.files("defs/", pattern = "soil_", full.names = TRUE)

input_rhessys = IOin_rhessys_input(
  version = rh_path,
  tec_file = "tecfiles/tec_daily",
  world_file = "CWWS32ssurgo.world.Y2018M11D1H1.state.Y2018M11D1H1.state",
  world_hdr_prefix = "CWWS32ssurgo",
  flowtable = "CWWS32ssurgo.flow",
  start = "2015 11 1 1",
  end = "2018 11 1 1",
  output_folder = "out",
  output_prefix = "cwws32valssurgo",
commandline_options = c(""))

input_tec_data = IOin_tec_std(start = "2017 11 1 1",
                              end = "2018 11 1 1",
                              output_state = FALSE)

input_hdr = IOin_hdr(
  basin = "defs/basin.def",
  hillslope = "defs/hillslope.def",
  zone = "defs/zone.def",
  soil = soildefs,
  landuse = "defs/lu_undev.def",
  stratum = c("defs/veg_deciduous/veg_deciduous.def","defs/veg_evergreen/veg_evergreen.def","defs/veg_deciduous_BES.def","defs/veg_eucalypt.def","defs/veg_grass.def","defs/veg_lawn_2cm.def","defs/veg_lawn_5cm.def","defs/veg_lawn_10cm.def","defs/veg_nonveg.def"),
  basestations = "clim/cwtws32local.base")

```
run SSURGO

```{r Run RHESSYS in Parallel using foreach SSURGO 1}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

n.cores <- parallel::detectCores() - 4  
  
my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
doParallel::registerDoParallel(cl = my.cluster)

start_time = Sys.time()
  
  foreach(i = 1:validationruns, .packages = 'RHESSysIOinR') %dopar% {
    
    stdpars<- IOin_cmd_pars(m = topssurgo[i,]$m,
                            k = topssurgo[i,]$k,
                            soil_dep= topssurgo[i,]$soil_dep,
                            m_v = topssurgo[i,]$m_v,
                            k_v = topssurgo[i,]$k_v,
                            gw1 = topssurgo[i,]$gw1,
                            gw2 = topssurgo[i,]$gw2,
                            pa = topssurgo[i,]$pa,
                            po = topssurgo[i,]$po,
                            vgseng1 = topssurgo[i,]$vgseng1,
                            vgseng2 = topssurgo[i,]$vgseng2,
                            vgseng3 = topssurgo[i,]$vgseng3)
    
    run_rhessys_single(input_rhessys = input_rhessys,
                       hdr_files = input_hdr,
                       tec_data = input_tec_data,
                       cmd_pars = stdpars,
                       output_filter = c(patchfilterpatch1,patchfilterpatch2,patchfilterpatch3),
                       runID = i)
  } 
  
  end_time = Sys.time()
  end_time - start_time
```
read 20 runs, create table and append to table for site 1

```{r}

setwd(system.file("extdata/", package = "RHESSysIOinR"))

getwd()


# Read in RHESSys Validation runs
for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch1_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valssurgo_run",i),temp_dataset)
}


##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valssurgo_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}

valmergesm1

validationsmNSElistsite1<- c()
validationsmlnNSElistsite1<- c()
validationsmKGElistsite1<- c()


#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site1"))))))
  validationsmKGElistsite1[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }


validationsmKGElistsite1

```
prepare to run model n times for patch 2 - only for validation time series

run model in parallel 

read 20 runs, create table and append to table for site 2

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

# Read in RHESSys Validation runs
for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch2_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valssurgo2_run",i),temp_dataset)
}


##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valssurgo2_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}


valmergesm1


validationsmNSElistsite2<- c()
validationsmlnNSElistsite2<- c()
validationsmKGElistsite2<- c()


#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site2"))))))
  validationsmKGElistsite2[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }


 validationsmKGElistsite2



```
prepare to run model 20 times for patch 3 - only for validation time series

run model in parallel

read 20 runs, create table and append to table for site 3

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

# Read in RHESSys Validation runs

for(i in 1:validationruns) { 
  temp_dataset<- read.csv(paste0("outputfilter/valws32patch3_",i,".csv"))
  temp_dataset$date<- as.Date(paste(temp_dataset$year, temp_dataset$month, temp_dataset$day, sep = "-"), format = "%Y-%m-%d")
  temp_dataset$rootzonevwc<- temp_dataset$rootzone.S*temp_dataset$rootzone.potential_sat 
  assign(paste0("cwws32valssurgo3_run",i),temp_dataset)
}


##merge calibration runs with observed so dates match up, instead of subsetting 

for(i in 1:validationruns) { assign(paste0("valmergesm",i), merge(obsws32smvalclean,eval(parse(text = paste0("cwws32valssurgo3_run",i))), by.x = "Date", by.y="date", all = FALSE))
  assign(paste0("valsubsetsm",i),eval(parse(text = paste0("valmergesm",i)))[eval(parse(text = paste0("valmergesm",i,"$date"))) >= Valdates[1] & eval(parse(text = paste0("valmergesm",i,"$date"))) <= Valdates[2], ])}


validationsmNSElistsite3<- c()
validationsmlnNSElistsite3<- c()
validationsmKGElistsite3<- c()

#for(i in 1:n) {  assign(paste0("vsmNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture"))))))
#  valsmNSElist[[i]]<- eval(parse(text = paste0("vsmNSEobs",i))) }

#for(i in 1:n) {  assign(paste0("vsmlnNSEobs",i), NSE(sim = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$rz_storage","/valsubsetsm",i,"$rootdepth")))), obs = as.numeric(eval(parse(text = paste0("valsubsetsm",i,"$mergedsoilmoisture")))), FUN = log, epsilon = "Pushpalatha2012", na.rm=TRUE))
#  valsmlnNSElist[[i]]<- eval(parse(text = paste0("vsmlnNSEobs",i))) }

for(i in 1:validationruns) {  assign(paste0("vsmKGEobs",i), KGE(sim = as.numeric(eval(parse(text = paste0("valmergesm",i,"$rootzonevwc")))), obs = as.numeric(eval(parse(text = paste0("valmergesm",i,"$Site3"))))))
  validationsmKGElistsite3[[i]]<- eval(parse(text = paste0("vsmKGEobs",i))) }


validationsmKGElistsite3

valmergesm2
 
```
merge all 3 sites

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

threesitevalidation1<- data.frame(cbind(site1=as.numeric(validationsmKGElistsite1),site2=as.numeric(validationsmKGElistsite2),site3=as.numeric(validationsmKGElistsite3)))

threesitevalidation<-threesitevalidation1

threesitevalidation$allsites<- rowMeans(threesitevalidation1,na.rm=FALSE)

threesitevalidation

boxplot(threesitevalidation1, main = "Boxplot of Patch Specific Root Zone Soil Moisture KGE - SSURGO")
boxplot(threesitevalidation, main = "Boxplot of Patch Specific Root Zone Soil Moisture KGE - SSURGO")

plot(validationsmKGElistsite1,validationsmKGElistsite3)

write.csv(threesitevalidation,'threesitevalidationssurgo.csv')


```
Group by group
color by site
plot kge on y axis
plot group on x axis

```{r}
setwd(system.file("extdata/", package = "RHESSysIOinR"))

ssurgovalidation<- read.csv("threesitevalidationssurgo.csv")
rssvalidation<- read.csv("threesitevalidationrss.csv")
staticvalidation<- read.csv("threesitevalidationstatic.csv")

staticvalidation$group <- 'Static'
rssvalidation$group <- 'RSS'
ssurgovalidation$group<- 'SSURGO'

combinedvalidation <- rbind(staticvalidation, rssvalidation, ssurgovalidation)


combinedvalidationlong1<- data.frame("KGE" = combinedvalidation$site1, "group" = combinedvalidation$group)
combinedvalidationlong2<- data.frame("KGE" = combinedvalidation$site2, "group" = combinedvalidation$group)
combinedvalidationlong3<- data.frame("KGE" = combinedvalidation$site3, "group" = combinedvalidation$group)
combinedvalidationlong4<- data.frame("KGE" = combinedvalidation$allsites, "group" = combinedvalidation$group)


combinedvalidationlong1$site = '1'
combinedvalidationlong2$site = '2'
combinedvalidationlong3$site = '3'
combinedvalidationlong4$site = 'All Sites'


combinedvalidationlong<- rbind(combinedvalidationlong1,combinedvalidationlong2, combinedvalidationlong3, combinedvalidationlong4)

combinedvalidationlong

#Remove NA values so that boxplot will work

validationnona<- subset(combinedvalidationlong,!is.na(combinedvalidationlong$KGE))


validationnona

threesitevalidationplot <- ggplot(validationnona, aes(group, y = KGE, fill = site)) + 
  geom_boxplot() + 
  ylab("Soil Moisture KGE - Validation") +
  xlab("Soil Map Input") +
  ggtitle("Patch Specific Soil Moisture KGE For ALL Calibration Runs, NA Excluded") +
  geom_hline(yintercept = -0.41, linetype = 2, col = 'RED') + 
  scale_y_continuous(breaks = seq(-6, 1, 1)) + 
  scale_fill_manual(values = c('#e41a1c', '#377eb8', '#4daf4a', '#984ea3')) + 
  theme_minimal()

threesitevalidationplot


```



combinedvalidation
